Split run-job and mise CLI responsibilities
Amp-Thread-ID: https://ampcode.com/threads/T-01a012c4-f3b3-76bd-81c2-aed4a1fddd7d
