+++ internal/cli/cli.go
@@ -23,7 +23,6 @@ import (
 	"runtime"
 	"slices"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
@@ -42,32 +41,23 @@ import (
 )
 
 const (
-	resultPublicationTimeout                    = 10 * time.Second
-	legacyRuntimeQueue                          = "hosted"
-	pluginConfigurationEnvironment              = "BUILDKITE_PLUGIN_CONFIGURATION"
-	pluginDevDarwinRuntimeEnvironment           = "BUILDKITE_GHA_PLUGIN_DEV_DARWIN_RUNTIME"
-	pluginDevLinuxRuntimeEnvironment            = "BUILDKITE_GHA_PLUGIN_DEV_LINUX_RUNTIME"
-	legacyTargetQueueEnvironment                = "BUILDKITE_GHA_TARGET_QUEUE"
-	legacyRuntimeImageEnvironment               = "BUILDKITE_GHA_RUNTIME_IMAGE"
-	repositoryProviderGitCredentialsEnvironment = "BUILDKITE_USE_REPOSITORY_PROVIDER_GIT_CREDENTIALS"
-	legacyGitHubAppGitCredentialsEnvironment    = "BUILDKITE_USE_GITHUB_APP_GIT_CREDENTIALS"
-	hostedProfile                               = "hosted"
-	legacyHostedTokenlessProfile                = "hosted-tokenless"
-	runtimeMiseArchiveDigest                    = "bd0930c0b619f51ddb60e32e5cce18a5533567b2f1ba9fc4875b9f39a2bb3ed8"
-	runtimeMiseBinaryDigest                     = "a238972a3162d710b85b28c324372e96ca4e4b486c81fe78695000d9fbc77c48"
-	runtimeMiseDarwinARM64ArchiveDigest         = "5b883c868a0748dd0c595d30fd000ec5138dfabdeef2c30222866ebf34af1ae3"
-	runtimeMiseDarwinARM64BinaryDigest          = "e777070540ffe22cf8b2b9f88aed88b461d0887d940c4f1c1a97359463cde6e1"
-	runtimeMiseArchiveLimit                     = 64 << 20
-	runtimeMiseBinaryLimit                      = 128 << 20
-	runtimeDistributionLimit                    = 256 << 20
-	pluginChecksumLimit                         = 4 << 20
-	pluginArchiveLimit                          = 256 << 20
-	maxWebhookMetadataBytes                     = 25 << 20
-	workflowCheckSummaryLimit                   = 65535
-	workflowCheckSummaryNotice                  = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
-	defaultNobleRunnerImage                     = "buildkite.namespace-images.com/agent-base@sha256:d8fa20ce298d9ee8673dc6e9ecddabb23ebbff3bcaf828d2011714a960cc5853"
-	defaultJammyRunnerImage                     = "buildkite.namespace-images.com/agent-base@sha256:3489d7d999869a11a243bdd9f91f318784cb74f0b57646c57424d10008b1e21d"
-	defaultMacOSRunnerQueue                     = "macos-medium"
+	legacyRuntimeQueue                = "hosted"
+	pluginConfigurationEnvironment    = "BUILDKITE_PLUGIN_CONFIGURATION"
+	pluginDevDarwinRuntimeEnvironment = "BUILDKITE_GHA_PLUGIN_DEV_DARWIN_RUNTIME"
+	pluginDevLinuxRuntimeEnvironment  = "BUILDKITE_GHA_PLUGIN_DEV_LINUX_RUNTIME"
+	legacyTargetQueueEnvironment      = "BUILDKITE_GHA_TARGET_QUEUE"
+	legacyRuntimeImageEnvironment     = "BUILDKITE_GHA_RUNTIME_IMAGE"
+	hostedProfile                     = "hosted"
+	legacyHostedTokenlessProfile      = "hosted-tokenless"
+	runtimeDistributionLimit          = 256 << 20
+	pluginChecksumLimit               = 4 << 20
+	pluginArchiveLimit                = 256 << 20
+	maxWebhookMetadataBytes           = 25 << 20
+	workflowCheckSummaryLimit         = 65535
+	workflowCheckSummaryNotice        = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
+	defaultNobleRunnerImage           = "buildkite.namespace-images.com/agent-base@sha256:d8fa20ce298d9ee8673dc6e9ecddabb23ebbff3bcaf828d2011714a960cc5853"
+	defaultJammyRunnerImage           = "buildkite.namespace-images.com/agent-base@sha256:3489d7d999869a11a243bdd9f91f318784cb74f0b57646c57424d10008b1e21d"
+	defaultMacOSRunnerQueue           = "macos-medium"
 )
 
 var runnerQueuePattern = regexp.MustCompile(`^[A-Za-z0-9_-]{1,255}$`)
@@ -76,10 +66,6 @@ var stableVersionPattern = regexp.MustCompile(`^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)
 
 var pluginReleaseBaseURL = "https://github.com/buildkite/buildkite-gha/releases/download"
 var pluginHTTPClient = securePluginHTTPClient()
-
-func repositoryProviderGitCredentialsEnabled(getenv func(string) string) bool {
-	return getenv(repositoryProviderGitCredentialsEnvironment) == "true" || getenv(legacyGitHubAppGitCredentialsEnvironment) == "true"
-}
 
 // Run executes the command and returns its process exit code.
 func Run(args []string, stdout, stderr io.Writer, version string) int {
@@ -384,923 +370,6 @@ func normalizePluginCommit(ctx context.Context, getenv func(string) string, sete
 	}
 	if err := setenv("BUILDKITE_COMMIT", commit); err != nil {
 		return fmt.Errorf("set resolved BUILDKITE_COMMIT: %w", err)
-	}
-	return nil
-}
-
-func runJob(args []string, stdout, stderr io.Writer, version, clientVersion string, agent transport.Agent) int {
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer stop()
-	return runJobContext(ctx, args, stdout, stderr, version, clientVersion, agent)
-}
-
-func runJobContext(ctx context.Context, args []string, stdout, stderr io.Writer, version, clientVersion string, agent transport.Agent) (code int) {
-	started := time.Now()
-	var result gharuntime.JobResult
-	details := &commandTelemetryDetails{}
-	defer func() {
-		outcome := telemetryOutcome(code, result.Conclusion, ctx.Err())
-		emitCommandTelemetry(telemetry.CommandRunJob, outcome, clientVersion, time.Since(started), details.forOutcome(outcome))
-	}()
-	options, err := runJobArgs(args)
-	if err != nil {
-		return usageError(stderr, "run-job: %v", err)
-	}
-	failureVisible := false
-	_, _ = fmt.Fprintln(stdout, "~~~ :package: Prepare GitHub Actions job")
-	defer func() {
-		if code != 0 && !failureVisible {
-			_, _ = fmt.Fprintln(stdout, "^^^ +++")
-		}
-	}()
-	if options.planProducer != "" {
-		planRoot, mkdirErr := os.MkdirTemp("", "buildkite-gha-plan-")
-		if mkdirErr != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: create plan directory: %v\n", mkdirErr)
-			return 1
-		}
-		defer func() { _ = os.RemoveAll(planRoot) }()
-		if err := agent.DownloadArtifact(ctx, options.planPath, planRoot, options.planProducer); err != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: download plan: %v\n", err)
-			return 1
-		}
-		options.planPath = filepath.Join(planRoot, filepath.FromSlash(options.planPath))
-	}
-	source, err := os.ReadFile(options.planPath)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	planDigest := transport.Digest(source)
-	expectedPlanDigest := options.planDigest
-	if expectedPlanDigest == "" {
-		expectedPlanDigest = os.Getenv("BUILDKITE_GHA_PLAN_DIGEST")
-	}
-	if expectedPlanDigest != "" {
-		if planDigest != expectedPlanDigest {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: plan digest %q does not match expected digest %q\n", planDigest, expectedPlanDigest)
-			return 1
-		}
-	}
-	job, err := plan.Decode(source)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	if job.Compiler.Version != version {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: plan compiler version %q does not match runtime version %q\n", job.Compiler.Version, version)
-		return 1
-	}
-	runtimeDigest, err := executableDigest()
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: verify runtime executable: %v\n", err)
-		return 1
-	}
-	if expected := job.RuntimeDistributionDigest(); runtimeDigest != expected {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: runtime distribution digest %q does not match plan digest %q\n", runtimeDigest, expected)
-		return 1
-	}
-	if err := verifyBuildkiteTarget(job); err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	if err := gharuntime.ValidateHost(job, runtime.GOOS, runtime.GOARCH); err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	producer, publish, err := resultProducer(job, planDigest, expectedPlanDigest)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	var artifactRoot string
-	if publish {
-		artifactRoot, err = os.MkdirTemp("", "buildkite-gha-results-")
-		if err != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: create result artifact root: %v\n", err)
-			return 1
-		}
-		defer func() { _ = os.RemoveAll(artifactRoot) }()
-	}
-	var actionMaterializer gharuntime.ActionMaterializer
-	if hasGitHubActionLocks(job.Actions) {
-		actionCache, err := os.MkdirTemp("", "buildkite-gha-actions-")
-		if err != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: create action cache: %v\n", err)
-			return 1
-		}
-		defer func() { _ = os.RemoveAll(actionCache) }()
-		store, err := actionsource.NewStore(actionCache, nil)
-		if err != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: configure action cache: %v\n", err)
-			return 1
-		}
-		actionMaterializer = store
-	}
-	var cacheCredentials gharuntime.CacheCredentialProvider
-	cacheRequired, err := cacheServiceRequired(job.Actions)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", err)
-		return 1
-	}
-	if cacheRequired || len(job.Actions) > 0 {
-		cacheCredentials, err = gharuntime.NewAgentCacheCredentials(gharuntime.AgentCacheConfig{
-			Endpoint:   os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-			JobID:      os.Getenv("BUILDKITE_JOB_ID"),
-			JobToken:   os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-			ResultsURL: os.Getenv("BUILDKITE_GHA_CACHE_URL"),
-		})
-		if err != nil && cacheRequired {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: configure actions/cache service: %v\n", err)
-			return 1
-		}
-		if err != nil {
-			cacheCredentials = nil
-		}
-	}
-	var workflowTokens gharuntime.WorkflowTokenProvider
-	if job.HasCapability("provider-token-write") {
-		githubTokens, tokenErr := gharuntime.NewAgentGitHubTokens(gharuntime.AgentGitHubTokenConfig{
-			Endpoint:         os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-			JobID:            os.Getenv("BUILDKITE_JOB_ID"),
-			JobToken:         os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-			OrganizationSlug: os.Getenv("BUILDKITE_ORGANIZATION_SLUG"),
-			PipelineSlug:     os.Getenv("BUILDKITE_PIPELINE_SLUG"),
-		})
-		if tokenErr != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: configure GitHub token service: %v\n", tokenErr)
-			return 1
-		}
-		workflowTokens = githubTokens
-	}
-	var oidcTokens gharuntime.OIDCTokenProvider
-	if job.IDTokenPermission == "write" {
-		config := gharuntime.AgentOIDCTokenConfig{
-			Endpoint: os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-			JobID:    os.Getenv("BUILDKITE_JOB_ID"),
-			JobToken: os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-		}
-		if job.OIDC != nil {
-			config.Claims = job.OIDC.Claims
-			config.AWSSessionTags = job.OIDC.AWSSessionTags
-			config.SubjectClaim = job.OIDC.SubjectClaim
-		}
-		oidcTokens, err = gharuntime.NewAgentOIDCTokens(config)
-		if err != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: configure OIDC token service: %v\n", err)
-			return 1
-		}
-	}
-	var repositoryCredentials *gharuntime.AgentRepositoryCredentials
-	if repositoryProviderGitCredentialsEnabled(os.Getenv) {
-		repositoryCredentials = &gharuntime.AgentRepositoryCredentials{
-			Agent:    os.Getenv("BUILDKITE_GHA_AGENT"),
-			Endpoint: os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-			JobID:    os.Getenv("BUILDKITE_JOB_ID"),
-			JobToken: os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-			NoHTTP2:  os.Getenv("BUILDKITE_NO_HTTP2"),
-		}
-	}
-	runnerToolCache := ""
-	if options.hostedToolCache {
-		runnerToolCache = buildkitepipeline.HostedToolCachePath
-	}
-	runner := gharuntime.Runner{
-		Stdout:      stdout,
-		Stderr:      stderr,
-		MiseDataDir: prepareMiseDataDir(os.Getenv("BUILDKITE_GHA_MISE_DATA_DIR"), stderr),
-		ToolCache:   runnerToolCache,
-		Docker:      os.Getenv("BUILDKITE_GHA_DOCKER"),
-		Git:         os.Getenv("BUILDKITE_GHA_GIT"),
-		Secrets: gharuntime.AgentSecrets{
-			Executable: os.Getenv("BUILDKITE_GHA_AGENT"),
-			Endpoint:   os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-			JobID:      os.Getenv("BUILDKITE_JOB_ID"),
-			JobToken:   os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-			NoHTTP2:    os.Getenv("BUILDKITE_NO_HTTP2"),
-		},
-		Redactor:              gharuntime.AgentRedactor{Executable: os.Getenv("BUILDKITE_GHA_AGENT")},
-		Actions:               actionMaterializer,
-		Artifacts:             agent,
-		Cache:                 cacheCredentials,
-		RepositoryCredentials: repositoryCredentials,
-		WorkflowToken:         workflowTokens,
-		OIDCToken:             oidcTokens,
-	}
-	runner.RuntimeExecutable, err = os.Executable()
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: resolve runtime executable: %v\n", err)
-		return 1
-	}
-	var privateRuntime string
-	if job.NeedsMise() {
-		runner.ResolveMise = func(ctx context.Context) (string, error) {
-			var err error
-			privateRuntime, err = os.MkdirTemp("", "buildkite-gha-runtime-")
-			if err != nil {
-				return "", fmt.Errorf("prepare action runtime: create private action runtime: %w", err)
-			}
-			mise, err := resolveRuntimeMise(ctx, os.Getenv("BUILDKITE_GHA_MISE"), runner.MiseDataDir, privateRuntime, stderr)
-			if err != nil {
-				return "", fmt.Errorf("prepare action runtime: %w", err)
-			}
-			return mise, nil
-		}
-		defer func() {
-			if privateRuntime != "" {
-				_ = os.RemoveAll(privateRuntime)
-			}
-		}()
-	}
-	var runErr error
-	if len(job.NeedSources) != 0 {
-		job.Needs, runErr = gharuntime.ResolveNeeds(ctx, agent, artifactRoot, producer.BuildID, job.NeedSources, job.NeedOutputs)
-		if runErr != nil {
-			details.setFailurePhase(telemetry.FailurePhaseSourceResolution)
-			runErr = fmt.Errorf("hydrate prerequisite results: %w", runErr)
-		}
-	}
-	if runErr == nil && len(job.CallGuards) != 0 {
-		job.CallGuards, runErr = gharuntime.ResolveCallGuards(ctx, agent, artifactRoot, producer.BuildID, job.CallGuards)
-		if runErr != nil {
-			details.setFailurePhase(telemetry.FailurePhaseSourceResolution)
-			runErr = fmt.Errorf("hydrate reusable-workflow call guards: %w", runErr)
-		}
-	}
-	if runErr == nil {
-		result, runErr = runner.RunJob(ctx, job, "")
-		failureVisible = result.FailureVisible()
-		if runErr != nil {
-			details.setFailurePhase(telemetry.FailurePhaseExecution)
-		}
-	}
-	if result.Conclusion == "" {
-		result.Conclusion = terminalErrorConclusion(ctx)
-	}
-	if options.resultPath != "" && result.Conclusion != "" {
-		if err := writeJobResult(options.resultPath, result); err != nil {
-			details.setFailurePhase(telemetry.FailurePhaseResultPublication)
-			runErr = errors.Join(runErr, err)
-			result.Conclusion = terminalErrorConclusion(ctx)
-		}
-	}
-	if publish && runErr != nil && !result.FailureVisible() {
-		_, _ = fmt.Fprintln(stdout, "+++ :warning: Prepare GitHub Actions job failed")
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", runErr)
-		failureVisible = true
-	}
-	if publish {
-		_, _ = fmt.Fprintln(stdout, "~~~ :package: Publish GitHub Actions result")
-		publication, err := publishTerminalResult(agent, artifactRoot, job, planDigest, producer, result)
-		if publication.MetadataMirrorError != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: result metadata mirror: %v\n", publication.MetadataMirrorError)
-		}
-		if publication.SummaryAnnotationError != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: job summary annotation: %v\n", publication.SummaryAnnotationError)
-		}
-		if publication.WarningAnnotationError != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: workflow warning annotation: %v\n", publication.WarningAnnotationError)
-		}
-		if publication.ErrorAnnotationError != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: workflow error annotation: %v\n", publication.ErrorAnnotationError)
-		}
-		if err != nil || publication.MetadataMirrorError != nil || publication.SummaryAnnotationError != nil || publication.WarningAnnotationError != nil || publication.ErrorAnnotationError != nil {
-			_, _ = fmt.Fprintln(stdout, "^^^ +++")
-			failureVisible = true
-		}
-		if err != nil {
-			details.setFailurePhase(telemetry.FailurePhaseResultPublication)
-			runErr = errors.Join(runErr, fmt.Errorf("publish terminal result: %w", err))
-		}
-	}
-	if runErr != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: %v\n", runErr)
-		if gharuntime.IsToleratedJobFailure(runErr) {
-			return buildkitepipeline.ContinueOnErrorExitStatus
-		}
-		return 1
-	}
-	return 0
-}
-
-func resolveRuntimeMise(ctx context.Context, configured, dataDir, privateRuntime string, stderr io.Writer) (string, error) {
-	return resolveRuntimeMiseWithInstaller(ctx, configured, dataDir, privateRuntime, stderr, installRuntimeMise)
-}
-
-func resolveRuntimeMiseWithInstaller(ctx context.Context, configured, dataDir, privateRuntime string, stderr io.Writer, install func(context.Context, string, string, io.Writer) (string, error)) (string, error) {
-	if configured != "" {
-		if !filepath.IsAbs(configured) {
-			return "", fmt.Errorf("BUILDKITE_GHA_MISE must be an absolute path")
-		}
-		return validateRuntimeMise(ctx, configured, "")
-	}
-	if candidate, err := exec.LookPath("mise"); err == nil {
-		if resolved, validationErr := validateRuntimeMise(ctx, candidate, ""); validationErr == nil {
-			return resolved, nil
-		}
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: mise on PATH is incompatible with minimum version %s; using the managed runtime copy\n", buildkitepipeline.MinimumMiseVersion)
-	}
-	return install(ctx, dataDir, privateRuntime, stderr)
-}
-
-func validateRuntimeMise(ctx context.Context, candidate, expectedDigest string) (string, error) {
-	resolved, err := validateRuntimeMiseFile(ctx, candidate, expectedDigest)
-	if err != nil {
-		return "", err
-	}
-	command := exec.CommandContext(ctx, resolved, "--version")
-	for _, value := range os.Environ() {
-		name, _, _ := strings.Cut(value, "=")
-		if !strings.HasPrefix(name, "MISE_") {
-			command.Env = append(command.Env, value)
-		}
-	}
-	output, err := command.Output()
-	if err != nil {
-		return "", fmt.Errorf("validate runtime mise executable: %w", err)
-	}
-	fields := strings.Fields(string(output))
-	if len(fields) == 0 {
-		return "", fmt.Errorf("validate runtime mise executable: empty version")
-	}
-	reported := fields[0]
-	if reported == "mise" && len(fields) > 1 {
-		reported = fields[1]
-	}
-	if !miseVersionAtLeast(reported, buildkitepipeline.MinimumMiseVersion) {
-		return "", fmt.Errorf("runtime mise executable reported version %q, want %q or newer", reported, buildkitepipeline.MinimumMiseVersion)
-	}
-	return resolved, nil
-}
-
-func validateRuntimeMiseFile(ctx context.Context, candidate, expectedDigest string) (string, error) {
-	if !filepath.IsAbs(candidate) {
-		absolute, err := filepath.Abs(candidate)
-		if err != nil {
-			return "", fmt.Errorf("resolve runtime mise path: %w", err)
-		}
-		candidate = absolute
-	}
-	resolved, err := filepath.EvalSymlinks(candidate)
-	if err != nil {
-		return "", fmt.Errorf("resolve runtime mise executable: %w", err)
-	}
-	info, err := os.Lstat(resolved)
-	if err != nil {
-		return "", fmt.Errorf("inspect runtime mise executable: %w", err)
-	}
-	if !info.Mode().IsRegular() || info.Mode().Perm()&0o111 == 0 {
-		return "", fmt.Errorf("runtime mise executable %q is not an executable regular file", resolved)
-	}
-	if expectedDigest != "" {
-		if info.Size() > runtimeMiseBinaryLimit {
-			return "", fmt.Errorf("runtime mise executable exceeds %d-byte limit", runtimeMiseBinaryLimit)
-		}
-		actual, err := fileSHA256(ctx, resolved, runtimeMiseBinaryLimit)
-		if err != nil {
-			return "", fmt.Errorf("hash runtime mise executable: %w", err)
-		}
-		if actual != expectedDigest {
-			return "", fmt.Errorf("runtime mise executable checksum mismatch")
-		}
-	}
-	return resolved, nil
-}
-
-func miseVersionAtLeast(actual, minimum string) bool {
-	parse := func(value string) ([3]int, bool) {
-		var parsed [3]int
-		parts := strings.Split(strings.TrimPrefix(value, "v"), ".")
-		if len(parts) != len(parsed) {
-			return parsed, false
-		}
-		for index, part := range parts {
-			number, err := strconv.Atoi(part)
-			if err != nil || number < 0 {
-				return parsed, false
-			}
-			parsed[index] = number
-		}
-		return parsed, true
-	}
-	got, ok := parse(actual)
-	if !ok {
-		return false
-	}
-	want, ok := parse(minimum)
-	if !ok {
-		return false
-	}
-	for index := range got {
-		if got[index] != want[index] {
-			return got[index] > want[index]
-		}
-	}
-	return true
-}
-
-type runtimeMiseRelease struct {
-	asset         string
-	cacheKey      string
-	archiveDigest string
-	binaryDigest  string
-}
-
-func selectRuntimeMiseRelease(goos, goarch string) (runtimeMiseRelease, error) {
-	switch goos + "/" + goarch {
-	case "linux/amd64":
-		return runtimeMiseRelease{
-			asset:         "linux-x64",
-			cacheKey:      "linux-amd64",
-			archiveDigest: runtimeMiseArchiveDigest,
-			binaryDigest:  runtimeMiseBinaryDigest,
-		}, nil
-	case "darwin/arm64":
-		return runtimeMiseRelease{
-			asset:         "macos-arm64",
-			cacheKey:      "darwin-arm64",
-			archiveDigest: runtimeMiseDarwinARM64ArchiveDigest,
-			binaryDigest:  runtimeMiseDarwinARM64BinaryDigest,
-		}, nil
-	default:
-		return runtimeMiseRelease{}, fmt.Errorf("managed mise is unavailable on %s/%s; set BUILDKITE_GHA_MISE to a compatible absolute path", goos, goarch)
-	}
-}
-
-func installRuntimeMise(ctx context.Context, dataDir, privateRuntime string, stderr io.Writer) (string, error) {
-	selected, err := selectRuntimeMiseRelease(runtime.GOOS, runtime.GOARCH)
-	if err != nil {
-		return "", err
-	}
-	root := dataDir
-	if root == "" {
-		root, err = os.UserCacheDir()
-		if err != nil {
-			return "", fmt.Errorf("resolve mise runtime cache: %w", err)
-		}
-		root = filepath.Join(root, "buildkite-gha", "mise", buildkitepipeline.MinimumMiseVersion)
-	} else {
-		root = filepath.Join(filepath.Dir(root), "runtime", buildkitepipeline.MinimumMiseVersion)
-	}
-	destination := filepath.Join(root, selected.cacheKey, "mise")
-	if resolved, err := validateRuntimeMiseFile(ctx, destination, selected.binaryDigest); err == nil {
-		return pinRuntimeMise(ctx, resolved, privateRuntime, selected.binaryDigest)
-	}
-	_, _ = fmt.Fprintf(stderr, "~~~ :mise: Install mise %s\n", buildkitepipeline.MinimumMiseVersion)
-	url := fmt.Sprintf("https://github.com/jdx/mise/releases/download/v%s/mise-v%s-%s.tar.gz", buildkitepipeline.MinimumMiseVersion, buildkitepipeline.MinimumMiseVersion, selected.asset)
-	client := &http.Client{
-		Timeout: 2 * time.Minute,
-		CheckRedirect: func(request *http.Request, via []*http.Request) error {
-			if request.URL.Scheme != "https" {
-				return errors.New("mise download redirected away from HTTPS")
-			}
-			if len(via) >= 10 {
-				return errors.New("too many mise download redirects")
-			}
-			return nil
-		},
-	}
-	cached, err := installRuntimeMiseFromPlatform(ctx, root, selected.cacheKey, client, url, selected.archiveDigest, selected.binaryDigest)
-	if err != nil {
-		return "", err
-	}
-	return pinRuntimeMise(ctx, cached, privateRuntime, selected.binaryDigest)
-}
-
-func pinRuntimeMise(ctx context.Context, cached, privateRuntime, expectedDigest string) (string, error) {
-	if privateRuntime == "" {
-		return "", fmt.Errorf("private action runtime directory is required")
-	}
-	resolvedRoot, err := canonicalNonSymlinkDirectory(privateRuntime)
-	if err != nil {
-		return "", fmt.Errorf("private action runtime directory contains a symlink")
-	}
-	source, err := os.Open(cached)
-	if err != nil {
-		return "", fmt.Errorf("open cached mise executable: %w", err)
-	}
-	defer func() { _ = source.Close() }()
-	destination := filepath.Join(resolvedRoot, "mise")
-	output, err := os.OpenFile(destination, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o500)
-	if err != nil {
-		return "", fmt.Errorf("create private mise executable: %w", err)
-	}
-	hash := sha256.New()
-	written, copyErr := io.Copy(io.MultiWriter(output, hash), io.LimitReader(source, runtimeMiseBinaryLimit+1))
-	closeErr := output.Close()
-	if copyErr != nil {
-		return "", fmt.Errorf("copy private mise executable: %w", copyErr)
-	}
-	if closeErr != nil {
-		return "", fmt.Errorf("write private mise executable: %w", closeErr)
-	}
-	if written > runtimeMiseBinaryLimit {
-		return "", fmt.Errorf("cached mise executable exceeds %d-byte limit", runtimeMiseBinaryLimit)
-	}
-	if hex.EncodeToString(hash.Sum(nil)) != expectedDigest {
-		return "", fmt.Errorf("cached mise executable checksum verification failed")
-	}
-	resolved, err := validateRuntimeMise(ctx, destination, expectedDigest)
-	if err != nil {
-		return "", fmt.Errorf("validate private mise executable: %w", err)
-	}
-	return resolved, nil
-}
-
-func installRuntimeMiseFrom(ctx context.Context, root string, client *http.Client, sourceURL, archiveDigest, binaryDigest string) (string, error) {
-	return installRuntimeMiseFromPlatform(ctx, root, "linux-x64", client, sourceURL, archiveDigest, binaryDigest)
-}
-
-func installRuntimeMiseFromPlatform(ctx context.Context, root, cacheKey string, client *http.Client, sourceURL, archiveDigest, binaryDigest string) (string, error) {
-	destinationDir := filepath.Join(root, cacheKey)
-	parent := filepath.Dir(destinationDir)
-	if err := os.MkdirAll(parent, 0o755); err != nil {
-		return "", fmt.Errorf("create mise runtime cache: %w", err)
-	}
-	resolvedParent, err := canonicalNonSymlinkDirectory(parent)
-	if err != nil {
-		return "", fmt.Errorf("mise runtime cache contains a symlink")
-	}
-	destinationDir = filepath.Join(resolvedParent, cacheKey)
-	destination := filepath.Join(destinationDir, "mise")
-	if resolved, err := validateRuntimeMiseFile(ctx, destination, binaryDigest); err == nil {
-		return resolved, nil
-	}
-	staging, err := os.MkdirTemp(resolvedParent, "."+cacheKey+".")
-	if err != nil {
-		return "", fmt.Errorf("stage mise runtime: %w", err)
-	}
-	defer func() { _ = os.RemoveAll(staging) }()
-	archive := filepath.Join(staging, "mise.tar.gz")
-	if err := downloadRuntimeMise(ctx, client, sourceURL, archive, archiveDigest); err != nil {
-		return "", err
-	}
-	stagedExecutable := filepath.Join(staging, "mise")
-	if err := extractRuntimeMise(archive, stagedExecutable, binaryDigest); err != nil {
-		return "", err
-	}
-	if _, err := validateRuntimeMiseFile(ctx, stagedExecutable, binaryDigest); err != nil {
-		return "", fmt.Errorf("validate downloaded mise executable: %w", err)
-	}
-	if err := os.Remove(archive); err != nil {
-		return "", fmt.Errorf("remove staged mise archive: %w", err)
-	}
-	if _, err := os.Lstat(destinationDir); err == nil {
-		if resolved, validationErr := validateRuntimeMiseFile(ctx, destination, binaryDigest); validationErr == nil {
-			return resolved, nil
-		}
-		invalid := destinationDir + fmt.Sprintf(".invalid-%d", time.Now().UnixNano())
-		if err := os.Rename(destinationDir, invalid); err == nil {
-			_ = os.RemoveAll(invalid)
-		}
-	} else if !errors.Is(err, os.ErrNotExist) {
-		return "", fmt.Errorf("inspect mise runtime cache: %w", err)
-	}
-	if err := os.Rename(staging, destinationDir); err != nil {
-		if resolved, validationErr := validateRuntimeMiseFile(ctx, destination, binaryDigest); validationErr == nil {
-			return resolved, nil
-		}
-		return "", fmt.Errorf("publish mise runtime cache: %w", err)
-	}
-	staging = ""
-	resolved, err := validateRuntimeMiseFile(ctx, destination, binaryDigest)
-	if err != nil {
-		return "", fmt.Errorf("validate installed mise cache: %w", err)
-	}
-	return resolved, nil
-}
-
-func downloadRuntimeMise(ctx context.Context, client *http.Client, sourceURL, destination, expectedDigest string) error {
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, sourceURL, nil)
-	if err != nil {
-		return fmt.Errorf("create mise download request: %w", err)
-	}
-	response, err := client.Do(request)
-	if err != nil {
-		return fmt.Errorf("download mise: %w", err)
-	}
-	defer func() { _ = response.Body.Close() }()
-	if response.StatusCode != http.StatusOK {
-		return fmt.Errorf("download mise: unexpected HTTP status %s", response.Status)
-	}
-	file, err := os.OpenFile(destination, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o600)
-	if err != nil {
-		return fmt.Errorf("create mise archive: %w", err)
-	}
-	hash := sha256.New()
-	written, copyErr := io.Copy(io.MultiWriter(file, hash), io.LimitReader(response.Body, runtimeMiseArchiveLimit+1))
-	closeErr := file.Close()
-	if copyErr != nil {
-		return fmt.Errorf("download mise archive: %w", copyErr)
-	}
-	if closeErr != nil {
-		return fmt.Errorf("write mise archive: %w", closeErr)
-	}
-	if written > runtimeMiseArchiveLimit {
-		return fmt.Errorf("download mise archive: exceeds %d-byte limit", runtimeMiseArchiveLimit)
-	}
-	if hex.EncodeToString(hash.Sum(nil)) != expectedDigest {
-		return fmt.Errorf("mise archive checksum verification failed")
-	}
-	return nil
-}
-
-func extractRuntimeMise(archive, destination, expectedDigest string) error {
-	file, err := os.Open(archive)
-	if err != nil {
-		return fmt.Errorf("open mise archive: %w", err)
-	}
-	defer func() { _ = file.Close() }()
-	gzipReader, err := gzip.NewReader(file)
-	if err != nil {
-		return fmt.Errorf("open mise gzip archive: %w", err)
-	}
-	defer func() { _ = gzipReader.Close() }()
-	tarReader := tar.NewReader(gzipReader)
-	found := false
-	for {
-		header, err := tarReader.Next()
-		if errors.Is(err, io.EOF) {
-			break
-		}
-		if err != nil {
-			return fmt.Errorf("read mise archive: %w", err)
-		}
-		if strings.TrimPrefix(filepath.ToSlash(header.Name), "./") != "mise/bin/mise" {
-			continue
-		}
-		if found {
-			return fmt.Errorf("mise archive contains duplicate executable")
-		}
-		if header.Typeflag != tar.TypeReg || header.Size <= 0 || header.Size > runtimeMiseBinaryLimit {
-			return fmt.Errorf("mise archive executable is not a bounded regular file")
-		}
-		output, err := os.OpenFile(destination, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o500)
-		if err != nil {
-			return fmt.Errorf("create mise executable: %w", err)
-		}
-		hash := sha256.New()
-		written, copyErr := io.Copy(io.MultiWriter(output, hash), tarReader)
-		closeErr := output.Close()
-		if copyErr != nil {
-			return fmt.Errorf("extract mise executable: %w", copyErr)
-		}
-		if closeErr != nil {
-			return fmt.Errorf("write mise executable: %w", closeErr)
-		}
-		if written != header.Size || hex.EncodeToString(hash.Sum(nil)) != expectedDigest {
-			return fmt.Errorf("mise executable checksum verification failed")
-		}
-		found = true
-	}
-	if !found {
-		return fmt.Errorf("mise archive does not contain mise/bin/mise")
-	}
-	return nil
-}
-
-func fileSHA256(ctx context.Context, path string, limit int64) (string, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return "", err
-	}
-	defer func() { _ = file.Close() }()
-	hash := sha256.New()
-	buffer := make([]byte, 32*1024)
-	var read int64
-	for read <= limit {
-		if err := ctx.Err(); err != nil {
-			return "", err
-		}
-		chunk := buffer
-		if remaining := limit + 1 - read; remaining < int64(len(chunk)) {
-			chunk = chunk[:remaining]
-		}
-		n, readErr := file.Read(chunk)
-		if n > 0 {
-			_, _ = hash.Write(chunk[:n])
-			read += int64(n)
-		}
-		if errors.Is(readErr, io.EOF) {
-			break
-		}
-		if readErr != nil {
-			return "", readErr
-		}
-	}
-	if read > limit {
-		return "", fmt.Errorf("exceeds %d-byte limit", limit)
-	}
-	return hex.EncodeToString(hash.Sum(nil)), nil
-}
-
-func prepareMiseDataDir(path string, stderr io.Writer) string {
-	if path == "" {
-		return ""
-	}
-	absolute, err := filepath.Abs(path)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: mise cache %q is invalid; using the ephemeral agent cache: %v\n", path, err)
-		return ""
-	}
-	if err := os.MkdirAll(absolute, 0o755); err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: mise cache %q is unavailable; using the ephemeral agent cache: %v\n", path, err)
-		return ""
-	}
-	resolved, err := canonicalNonSymlinkDirectory(absolute)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: run-job: warning: mise cache %q is not a real directory; using the ephemeral agent cache\n", path)
-		return ""
-	}
-	return resolved
-}
-
-func canonicalNonSymlinkDirectory(path string) (string, error) {
-	absolute, err := filepath.Abs(path)
-	if err != nil {
-		return "", err
-	}
-	info, err := os.Lstat(absolute)
-	if err != nil {
-		return "", err
-	}
-	if !info.IsDir() || info.Mode()&os.ModeSymlink != 0 {
-		return "", fmt.Errorf("not a non-symlink directory")
-	}
-	resolved, err := filepath.EvalSymlinks(absolute)
-	if err != nil {
-		return "", err
-	}
-	canonicalInfo, err := os.Stat(resolved)
-	if err != nil || !os.SameFile(info, canonicalInfo) {
-		return "", fmt.Errorf("directory changed while canonicalizing")
-	}
-	return resolved, nil
-}
-
-func hasGitHubActionLocks(locks []plan.ActionLock) bool {
-	for _, lock := range locks {
-		if lock.Source == "github" {
-			return true
-		}
-	}
-	return false
-}
-
-func cacheServiceRequired(locks []plan.ActionLock) (bool, error) {
-	required := false
-	for _, lock := range locks {
-		identity := actionintegration.Identity{Source: lock.Source, Repository: lock.Repository, Path: lock.Path}
-		descriptor, _ := actionintegration.Lookup(identity)
-		if descriptor.Service == actionintegration.ServiceCache {
-			required = true
-			if _, _, err := actionintegration.Admit(identity, lock.Commit); err != nil {
-				return false, fmt.Errorf("unsupported cache action: %w", err)
-			}
-		}
-	}
-	return required, nil
-}
-
-func resultProducer(job plan.Job, planDigest, expectedDigest string) (transport.Producer, bool, error) {
-	if os.Getenv("BUILDKITE") == "" {
-		if len(job.NeedSources) != 0 || callGuardsHaveNeedSources(job.CallGuards) {
-			return transport.Producer{}, false, fmt.Errorf("plans with prerequisites require Buildkite result identity")
-		}
-		return transport.Producer{}, false, nil
-	}
-	if expectedDigest == "" {
-		return transport.Producer{}, false, fmt.Errorf("result publication in Buildkite requires BUILDKITE_GHA_PLAN_DIGEST")
-	}
-	if expectedDigest != planDigest {
-		return transport.Producer{}, false, fmt.Errorf("plan digest %q does not match expected digest %q", planDigest, expectedDigest)
-	}
-	producer := transport.Producer{
-		BuildID: os.Getenv("BUILDKITE_BUILD_ID"),
-		JobID:   os.Getenv("BUILDKITE_JOB_ID"),
-		StepKey: os.Getenv("BUILDKITE_STEP_KEY"),
-	}
-	if err := producer.Validate(); err != nil {
-		return transport.Producer{}, false, fmt.Errorf("result publication in Buildkite requires valid BUILDKITE_BUILD_ID, BUILDKITE_JOB_ID, and BUILDKITE_STEP_KEY: %w", err)
-	}
-	if producer.StepKey != job.Target.StepKey {
-		return transport.Producer{}, false, fmt.Errorf("result producer step %q does not match plan target %q", producer.StepKey, job.Target.StepKey)
-	}
-	return producer, true, nil
-}
-
-func callGuardsHaveNeedSources(guards []plan.CallGuard) bool {
-	for _, guard := range guards {
-		if len(guard.NeedSources) != 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func writeJobResult(path string, result gharuntime.JobResult) error {
-	encoded, err := json.MarshalIndent(result, "", "  ")
-	if err != nil {
-		return fmt.Errorf("encode result: %w", err)
-	}
-	encoded = append(encoded, '\n')
-	if err := os.WriteFile(path, encoded, 0o600); err != nil {
-		return fmt.Errorf("write result: %w", err)
-	}
-	return nil
-}
-
-func terminalErrorConclusion(ctx context.Context) string {
-	if ctx.Err() != nil {
-		return "cancelled"
-	}
-	return "failure"
-}
-
-func publishTerminalResult(agent transport.Agent, root string, job plan.Job, planDigest string, producer transport.Producer, result gharuntime.JobResult) (transport.Publication, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), resultPublicationTimeout)
-	defer cancel()
-	workflow := strings.TrimPrefix(job.Workflow.Digest, "sha256:")
-	return gharuntime.PublishJobResult(ctx, agent, root, workflow, job.Target.StepKey, planDigest, producer, result)
-}
-
-type runJobOptions struct {
-	planPath        string
-	planDigest      string
-	planProducer    string
-	resultPath      string
-	hostedToolCache bool
-}
-
-func runJobArgs(args []string) (runJobOptions, error) {
-	var options runJobOptions
-	seen := map[string]bool{}
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "--hosted-tool-cache":
-			if seen[args[i]] {
-				return runJobOptions{}, fmt.Errorf("%s may only be specified once", args[i])
-			}
-			seen[args[i]] = true
-			options.hostedToolCache = true
-		case "--plan", "--plan-digest", "--plan-producer", "--result":
-			option := args[i]
-			if seen[option] {
-				return runJobOptions{}, fmt.Errorf("%s may only be specified once", option)
-			}
-			seen[option] = true
-			i++
-			if i == len(args) {
-				return runJobOptions{}, fmt.Errorf("%s requires a value", option)
-			}
-			switch option {
-			case "--plan":
-				options.planPath = args[i]
-			case "--plan-digest":
-				options.planDigest = args[i]
-			case "--plan-producer":
-				options.planProducer = args[i]
-			case "--result":
-				options.resultPath = args[i]
-			}
-		default:
-			return runJobOptions{}, fmt.Errorf("unknown option %q", args[i])
-		}
-	}
-	if seen["--plan"] {
-		if seen["--plan-digest"] || seen["--plan-producer"] {
-			return runJobOptions{}, fmt.Errorf("--plan cannot be combined with --plan-digest or --plan-producer")
-		}
-		if options.planPath == "" {
-			return runJobOptions{}, fmt.Errorf("--plan requires a path")
-		}
-		return options, nil
-	}
-	if options.planDigest == "" || options.planProducer == "" {
-		return runJobOptions{}, fmt.Errorf("--plan or both --plan-digest and --plan-producer are required")
-	}
-	planPath, err := buildkitepipeline.PlanPath(options.planDigest)
-	if err != nil {
-		return runJobOptions{}, err
-	}
-	options.planPath = planPath
-	return options, nil
-}
-
-func verifyBuildkiteTarget(job plan.Job) error {
-	stepKey := os.Getenv("BUILDKITE_STEP_KEY")
-	queue := os.Getenv("BUILDKITE_AGENT_META_DATA_QUEUE")
-	if os.Getenv("BUILDKITE") != "" && stepKey == "" {
-		return fmt.Errorf("buildkite execution requires BUILDKITE_STEP_KEY")
-	}
-	if os.Getenv("BUILDKITE") != "" && job.Target.Queue != "" && queue == "" {
-		return fmt.Errorf("buildkite execution with an explicit target queue requires BUILDKITE_AGENT_META_DATA_QUEUE")
-	}
-	if stepKey != "" && stepKey != job.Target.StepKey {
-		return fmt.Errorf("plan targets step %q, executing step is %q", job.Target.StepKey, stepKey)
-	}
-	if job.Target.Queue != "" && queue != "" && queue != job.Target.Queue {
-		return fmt.Errorf("plan targets queue %q, executing queue is %q", job.Target.Queue, queue)
 	}
 	return nil
 }
