+++ internal/cli/cli_test.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"crypto/sha256"
 	"encoding/binary"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -31,7 +30,6 @@ import (
 	"github.com/buildkite/buildkite-gha/internal/compatibility"
 	"github.com/buildkite/buildkite-gha/internal/compiler"
 	"github.com/buildkite/buildkite-gha/internal/plan"
-	gharuntime "github.com/buildkite/buildkite-gha/internal/runtime"
 	"github.com/buildkite/buildkite-gha/internal/telemetry"
 	"github.com/buildkite/buildkite-gha/internal/transport"
 	"go.yaml.in/yaml/v4"
@@ -6231,50 +6229,6 @@ func TestRunUploadJavaScriptActionRequiresRuntimeMiseWithoutTransport(t *testing
 	}
 }
 
-func TestPrepareMiseDataDirFallsBackWhenCacheIsUnavailable(t *testing.T) {
-	var stderr bytes.Buffer
-	root, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal(err)
-	}
-	dir := filepath.Join(root, "mise")
-	if got := prepareMiseDataDir(dir, &stderr); got != dir || stderr.Len() != 0 {
-		t.Fatalf("prepareMiseDataDir() = %q, stderr = %q", got, stderr.String())
-	}
-
-	file := filepath.Join(t.TempDir(), "not-a-directory")
-	if err := os.WriteFile(file, []byte("x"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	stderr.Reset()
-	if got := prepareMiseDataDir(file, &stderr); got != "" || !strings.Contains(stderr.String(), "using the ephemeral agent cache") {
-		t.Fatalf("prepareMiseDataDir(file) = %q, stderr = %q", got, stderr.String())
-	}
-
-	parent, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal(err)
-	}
-	target, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal(err)
-	}
-	link := filepath.Join(parent, "linked-cache")
-	if err := os.Symlink(target, link); err != nil {
-		t.Fatal(err)
-	}
-	stderr.Reset()
-	want := filepath.Join(target, "mise")
-	if got := prepareMiseDataDir(filepath.Join(link, "mise"), &stderr); got != want || stderr.Len() != 0 {
-		t.Fatalf("prepareMiseDataDir(aliased ancestor) = %q, stderr = %q, want %q", got, stderr.String(), want)
-	}
-
-	stderr.Reset()
-	if got := prepareMiseDataDir(link, &stderr); got != "" || !strings.Contains(stderr.String(), "not a real directory") {
-		t.Fatalf("prepareMiseDataDir(symlink root) = %q, stderr = %q", got, stderr.String())
-	}
-}
-
 func TestRunUploadAllowsCompilerVerifiedLocalDockerfileAction(t *testing.T) {
 	requireImporterHost(t)
 	root := t.TempDir()
@@ -6328,469 +6282,6 @@ func writeFakeNode(t *testing.T, root string, major int) string {
 		t.Fatal(err)
 	}
 	return path
-}
-
-func setFakeMise(t *testing.T, version string) string {
-	t.Helper()
-	root := canonicalTempDir(t)
-	mise := filepath.Join(root, "mise")
-	script := "#!/bin/sh\nif [ -n \"${MISE_TEST_POISON:-}\" ]; then printf 'poisoned\\n'; else printf '" + version + " linux-x64 (test)\\n'; fi\n"
-	if err := os.WriteFile(mise, []byte(script), 0o700); err != nil {
-		t.Fatal(err)
-	}
-	t.Setenv("PATH", root+string(os.PathListSeparator)+os.Getenv("PATH"))
-	return mise
-}
-
-func TestResolveRuntimeMisePinsRequiredExecutable(t *testing.T) {
-	realMise := setFakeMise(t, buildkitepipeline.MinimumMiseVersion)
-	linkRoot := t.TempDir()
-	link := filepath.Join(linkRoot, "mise")
-	if err := os.Symlink(realMise, link); err != nil {
-		t.Fatal(err)
-	}
-	t.Setenv("PATH", linkRoot)
-	t.Setenv("MISE_TEST_POISON", "must-not-reach-version-check")
-	got, err := resolveRuntimeMise(context.Background(), "", t.TempDir(), t.TempDir(), io.Discard)
-	if err != nil {
-		t.Fatal(err)
-	}
-	want, err := filepath.EvalSymlinks(realMise)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got != want || !filepath.IsAbs(got) {
-		t.Fatalf("resolveRuntimeMise() = %q, want pinned %q", got, want)
-	}
-}
-
-func TestResolveRuntimeMiseAcceptsNewerVersion(t *testing.T) {
-	realMise := setFakeMise(t, "2026.8.1")
-	got, err := resolveRuntimeMiseWithInstaller(context.Background(), "", t.TempDir(), t.TempDir(), io.Discard, func(context.Context, string, string, io.Writer) (string, error) {
-		return "", errors.New("unexpected managed mise install")
-	})
-	if err != nil || got != realMise {
-		t.Fatalf("resolveRuntimeMiseWithInstaller() = %q, %v; want %q", got, err, realMise)
-	}
-}
-
-func TestResolveRuntimeMiseAcceptsPrefixedVersionOutput(t *testing.T) {
-	root := canonicalTempDir(t)
-	mise := filepath.Join(root, "mise")
-	if err := os.WriteFile(mise, []byte("#!/bin/sh\nprintf 'mise v2026.8.1 linux-x64 (test)\\n'\n"), 0o700); err != nil {
-		t.Fatal(err)
-	}
-	got, err := resolveRuntimeMise(context.Background(), mise, t.TempDir(), t.TempDir(), io.Discard)
-	if err != nil || got != mise {
-		t.Fatalf("resolveRuntimeMise() = %q, %v; want %q", got, err, mise)
-	}
-}
-
-func TestMiseVersionAtLeast(t *testing.T) {
-	for _, test := range []struct {
-		actual string
-		want   bool
-	}{
-		{actual: "2026.5.12", want: true},
-		{actual: "v2026.5.12", want: true},
-		{actual: "2026.5.13", want: true},
-		{actual: "2026.8.1", want: true},
-		{actual: "2027.1.1", want: true},
-		{actual: "2026.5.11"},
-		{actual: "2025.12.31"},
-		{actual: "2026.5"},
-		{actual: "latest"},
-	} {
-		if got := miseVersionAtLeast(test.actual, buildkitepipeline.MinimumMiseVersion); got != test.want {
-			t.Errorf("miseVersionAtLeast(%q, %q) = %t, want %t", test.actual, buildkitepipeline.MinimumMiseVersion, got, test.want)
-		}
-	}
-}
-
-func TestResolveRuntimeMiseInstallsManagedCopyWhenNeeded(t *testing.T) {
-	for _, test := range []struct {
-		name    string
-		version string
-	}{
-		{name: "missing"},
-		{name: "old PATH version", version: "2026.5.11"},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			pathRoot := t.TempDir()
-			if test.version != "" {
-				mise := filepath.Join(pathRoot, "mise")
-				if err := os.WriteFile(mise, []byte("#!/bin/sh\nprintf '"+test.version+" linux-x64 (test)\\n'\n"), 0o700); err != nil {
-					t.Fatal(err)
-				}
-			}
-			t.Setenv("PATH", pathRoot)
-			dataDir := t.TempDir()
-			managed := filepath.Join(t.TempDir(), "mise")
-			called := 0
-			privateRuntime := t.TempDir()
-			installer := func(_ context.Context, gotDataDir, gotPrivateRuntime string, _ io.Writer) (string, error) {
-				called++
-				if gotDataDir != dataDir {
-					t.Fatalf("installer data dir = %q, want %q", gotDataDir, dataDir)
-				}
-				if gotPrivateRuntime != privateRuntime {
-					t.Fatalf("installer private runtime = %q, want %q", gotPrivateRuntime, privateRuntime)
-				}
-				return managed, nil
-			}
-			got, err := resolveRuntimeMiseWithInstaller(context.Background(), "", dataDir, privateRuntime, io.Discard, installer)
-			if err != nil || got != managed || called != 1 {
-				t.Fatalf("resolveRuntimeMiseWithInstaller() = %q, %v; calls = %d", got, err, called)
-			}
-		})
-	}
-}
-
-func TestResolveRuntimeMiseRejectsInvalidExplicitOverride(t *testing.T) {
-	t.Run("configured path must be absolute", func(t *testing.T) {
-		if _, err := resolveRuntimeMise(context.Background(), "mise", t.TempDir(), t.TempDir(), io.Discard); err == nil || !strings.Contains(err.Error(), "must be an absolute path") {
-			t.Fatalf("resolveRuntimeMise() error = %v", err)
-		}
-	})
-	t.Run("old version", func(t *testing.T) {
-		mise := setFakeMise(t, "2026.5.11")
-		if _, err := resolveRuntimeMise(context.Background(), mise, t.TempDir(), t.TempDir(), io.Discard); err == nil || !strings.Contains(err.Error(), `reported version "2026.5.11", want "2026.5.12" or newer`) {
-			t.Fatalf("resolveRuntimeMise() error = %v", err)
-		}
-	})
-	t.Run("not executable", func(t *testing.T) {
-		mise := filepath.Join(t.TempDir(), "mise")
-		if err := os.WriteFile(mise, []byte("mise"), 0o600); err != nil {
-			t.Fatal(err)
-		}
-		if _, err := resolveRuntimeMise(context.Background(), mise, t.TempDir(), t.TempDir(), io.Discard); err == nil || !strings.Contains(err.Error(), "not an executable regular file") {
-			t.Fatalf("resolveRuntimeMise() error = %v", err)
-		}
-	})
-}
-
-func TestInstallRuntimeMiseDownloadsVerifiesAndReusesCache(t *testing.T) {
-	binary := []byte("#!/bin/sh\nprintf '" + buildkitepipeline.MinimumMiseVersion + " linux-x64 (test)\\n'\n")
-	archive := runtimeMiseTestArchive(t, binary)
-	archiveHash := sha256.Sum256(archive)
-	binaryHash := sha256.Sum256(binary)
-	requests := 0
-	server := httptest.NewServer(http.HandlerFunc(func(response http.ResponseWriter, _ *http.Request) {
-		requests++
-		_, _ = response.Write(archive)
-	}))
-	defer server.Close()
-	base, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal(err)
-	}
-	realParent := filepath.Join(base, "real")
-	if err := os.Mkdir(realParent, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	logicalParent := filepath.Join(base, "logical")
-	if err := os.Symlink(realParent, logicalParent); err != nil {
-		t.Skipf("symlinks unsupported: %v", err)
-	}
-	root := filepath.Join(logicalParent, "cache")
-	want := filepath.Join(realParent, "cache", "linux-x64", "mise")
-	for i := 0; i < 2; i++ {
-		got, err := installRuntimeMiseFrom(context.Background(), root, server.Client(), server.URL, hex.EncodeToString(archiveHash[:]), hex.EncodeToString(binaryHash[:]))
-		if err != nil || got != want {
-			t.Fatalf("installRuntimeMiseFrom() = %q, %v", got, err)
-		}
-	}
-	if requests != 1 {
-		t.Fatalf("mise archive requests = %d, want one cache miss", requests)
-	}
-	if got, err := os.ReadFile(want); err != nil || !bytes.Equal(got, binary) {
-		t.Fatalf("installed mise = %q, %v", got, err)
-	}
-}
-
-func TestInstallRuntimeMiseRejectsInvalidArchive(t *testing.T) {
-	binary := []byte("#!/bin/sh\nprintf '" + buildkitepipeline.MinimumMiseVersion + " linux-x64 (test)\\n'\n")
-	archive := runtimeMiseTestArchive(t, binary)
-	server := httptest.NewServer(http.HandlerFunc(func(response http.ResponseWriter, _ *http.Request) {
-		_, _ = response.Write(archive)
-	}))
-	defer server.Close()
-	binaryHash := sha256.Sum256(binary)
-	if _, err := installRuntimeMiseFrom(context.Background(), t.TempDir(), server.Client(), server.URL, strings.Repeat("0", 64), hex.EncodeToString(binaryHash[:])); err == nil || !strings.Contains(err.Error(), "archive checksum") {
-		t.Fatalf("installRuntimeMiseFrom() error = %v", err)
-	}
-}
-
-func TestValidateRuntimeMiseRejectsOversizedCacheEntry(t *testing.T) {
-	cached := filepath.Join(t.TempDir(), "mise")
-	file, err := os.OpenFile(cached, os.O_CREATE|os.O_WRONLY, 0o500)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := file.Truncate(runtimeMiseBinaryLimit + 1); err != nil {
-		_ = file.Close()
-		t.Fatal(err)
-	}
-	if err := file.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if _, err := validateRuntimeMiseFile(context.Background(), cached, strings.Repeat("0", 64)); err == nil || !strings.Contains(err.Error(), "exceeds") {
-		t.Fatalf("validateRuntimeMiseFile() error = %v, want size rejection", err)
-	}
-}
-
-func TestManagedMiseCacheIsNotExecutedBeforePrivateCopy(t *testing.T) {
-	root := canonicalTempDir(t)
-	marker := filepath.Join(t.TempDir(), "executed")
-	binary := []byte("#!/bin/sh\nprintf ran > '" + marker + "'\nprintf '" + buildkitepipeline.MinimumMiseVersion + " linux-x64 (test)\\n'\n")
-	digest := sha256.Sum256(binary)
-	cached := filepath.Join(root, "linux-x64", "mise")
-	if err := os.MkdirAll(filepath.Dir(cached), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(cached, binary, 0o500); err != nil {
-		t.Fatal(err)
-	}
-	got, err := installRuntimeMiseFrom(context.Background(), root, nil, "", "", hex.EncodeToString(digest[:]))
-	if err != nil || got != cached {
-		t.Fatalf("installRuntimeMiseFrom() = %q, %v; want cache hit %q", got, err, cached)
-	}
-	if _, err := os.Stat(marker); !errors.Is(err, os.ErrNotExist) {
-		t.Fatalf("shared cache executable ran during validation: %v", err)
-	}
-	if _, err := pinRuntimeMise(context.Background(), cached, t.TempDir(), hex.EncodeToString(digest[:])); err != nil {
-		t.Fatal(err)
-	}
-	if _, err := os.Stat(marker); err != nil {
-		t.Fatalf("private executable did not run during version validation: %v", err)
-	}
-}
-
-func TestManagedMiseColdCacheIsNotExecutedBeforePrivateCopy(t *testing.T) {
-	marker := filepath.Join(t.TempDir(), "executed")
-	binary := []byte("#!/bin/sh\nprintf ran > '" + marker + "'\nprintf '" + buildkitepipeline.MinimumMiseVersion + " linux-x64 (test)\\n'\n")
-	archive := runtimeMiseTestArchive(t, binary)
-	archiveDigest := sha256.Sum256(archive)
-	binaryDigest := sha256.Sum256(binary)
-	server := httptest.NewServer(http.HandlerFunc(func(response http.ResponseWriter, _ *http.Request) {
-		_, _ = response.Write(archive)
-	}))
-	defer server.Close()
-	cached, err := installRuntimeMiseFrom(context.Background(), t.TempDir(), server.Client(), server.URL, hex.EncodeToString(archiveDigest[:]), hex.EncodeToString(binaryDigest[:]))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := os.Stat(marker); !errors.Is(err, os.ErrNotExist) {
-		t.Fatalf("shared staging executable ran during validation: %v", err)
-	}
-	if _, err := pinRuntimeMise(context.Background(), cached, t.TempDir(), hex.EncodeToString(binaryDigest[:])); err != nil {
-		t.Fatal(err)
-	}
-	if _, err := os.Stat(marker); err != nil {
-		t.Fatalf("private executable did not run during version validation: %v", err)
-	}
-}
-
-func TestPinRuntimeMiseCopiesVerifiedBytesPrivately(t *testing.T) {
-	binary := []byte("#!/bin/sh\nprintf '" + buildkitepipeline.MinimumMiseVersion + " linux-x64 (test)\\n'\n")
-	digest := sha256.Sum256(binary)
-	cached := filepath.Join(t.TempDir(), "mise")
-	if err := os.WriteFile(cached, binary, 0o500); err != nil {
-		t.Fatal(err)
-	}
-	base, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal(err)
-	}
-	realParent := filepath.Join(base, "real")
-	if err := os.Mkdir(realParent, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	logicalParent := filepath.Join(base, "logical")
-	if err := os.Symlink(realParent, logicalParent); err != nil {
-		t.Skipf("symlinks unsupported: %v", err)
-	}
-	privateRuntime := filepath.Join(logicalParent, "runtime")
-	if err := os.Mkdir(privateRuntime, 0o700); err != nil {
-		t.Fatal(err)
-	}
-	got, err := pinRuntimeMise(context.Background(), cached, privateRuntime, hex.EncodeToString(digest[:]))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got != filepath.Join(realParent, "runtime", "mise") {
-		t.Fatalf("pinRuntimeMise() = %q, want private executable", got)
-	}
-	if err := os.Chmod(cached, 0o700); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(cached, []byte("tampered"), 0o700); err != nil {
-		t.Fatal(err)
-	}
-	if copied, err := os.ReadFile(got); err != nil || !bytes.Equal(copied, binary) {
-		t.Fatalf("private mise changed with cache: %q, %v", copied, err)
-	}
-	if _, err := pinRuntimeMise(context.Background(), cached, t.TempDir(), hex.EncodeToString(digest[:])); err == nil || !strings.Contains(err.Error(), "checksum") {
-		t.Fatalf("pinRuntimeMise() accepted tampered cache: %v", err)
-	}
-	linkedRuntime := filepath.Join(base, "linked-runtime")
-	if err := os.Symlink(filepath.Join(realParent, "runtime"), linkedRuntime); err != nil {
-		t.Fatal(err)
-	}
-	if _, err := pinRuntimeMise(context.Background(), got, linkedRuntime, hex.EncodeToString(digest[:])); err == nil || !strings.Contains(err.Error(), "contains a symlink") {
-		t.Fatalf("pinRuntimeMise() accepted symlink root: %v", err)
-	}
-}
-
-func TestInstallRuntimeMiseLiveRelease(t *testing.T) {
-	if os.Getenv("BUILDKITE_GHA_LIVE_REQUIRED") != "1" {
-		t.Skip("set BUILDKITE_GHA_LIVE_REQUIRED=1 to verify the pinned mise release")
-	}
-	privateRuntime := t.TempDir()
-	got, err := installRuntimeMise(context.Background(), t.TempDir(), privateRuntime, io.Discard)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if filepath.Dir(got) != privateRuntime {
-		t.Fatalf("installed mise path = %q, want job-private root %q", got, privateRuntime)
-	}
-	selected, err := selectRuntimeMiseRelease(runtime.GOOS, runtime.GOARCH)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := validateRuntimeMise(context.Background(), got, selected.binaryDigest); err != nil {
-		t.Fatalf("validate installed mise release: %v", err)
-	}
-}
-
-func TestSelectRuntimeMiseRelease(t *testing.T) {
-	for _, test := range []struct {
-		goos, goarch, asset, cacheKey, archiveDigest, binaryDigest string
-	}{
-		{"linux", "amd64", "linux-x64", "linux-amd64", runtimeMiseArchiveDigest, runtimeMiseBinaryDigest},
-		{"darwin", "arm64", "macos-arm64", "darwin-arm64", runtimeMiseDarwinARM64ArchiveDigest, runtimeMiseDarwinARM64BinaryDigest},
-	} {
-		selected, err := selectRuntimeMiseRelease(test.goos, test.goarch)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if selected.asset != test.asset || selected.cacheKey != test.cacheKey || selected.archiveDigest != test.archiveDigest || selected.binaryDigest != test.binaryDigest {
-			t.Fatalf("selectRuntimeMiseRelease(%s/%s) = %#v", test.goos, test.goarch, selected)
-		}
-	}
-	if _, err := selectRuntimeMiseRelease("linux", "arm64"); err == nil || !strings.Contains(err.Error(), "unavailable") {
-		t.Fatalf("selectRuntimeMiseRelease() unsupported platform error = %v", err)
-	}
-}
-
-func runtimeMiseTestArchive(t *testing.T, binary []byte) []byte {
-	t.Helper()
-	var out bytes.Buffer
-	gzipWriter := gzip.NewWriter(&out)
-	tarWriter := tar.NewWriter(gzipWriter)
-	if err := tarWriter.WriteHeader(&tar.Header{Name: "mise/bin/mise", Mode: 0o755, Size: int64(len(binary)), Typeflag: tar.TypeReg}); err != nil {
-		t.Fatal(err)
-	}
-	if _, err := tarWriter.Write(binary); err != nil {
-		t.Fatal(err)
-	}
-	if err := tarWriter.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if err := gzipWriter.Close(); err != nil {
-		t.Fatal(err)
-	}
-	return out.Bytes()
-}
-
-func TestRunJobPublishesFailureWhenExplicitRuntimeMiseIsInvalid(t *testing.T) {
-	job := cliRunJobPlan()
-	job.Schema = plan.Schema
-	requiresMise := true
-	job.RequiresMise = &requiresMise
-	job.Actions = []plan.ActionLock{{
-		ID: "a-0000000000000001", Source: "workspace", Path: "actions/build",
-		SourceDigest: "sha256:" + strings.Repeat("a", 64),
-	}}
-	job.Steps = []plan.Step{{
-		ID: "local", Kind: "uses", Uses: "./actions/build",
-		Action: &plan.ActionSelector{Lock: "a-0000000000000001"},
-	}}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	t.Setenv("BUILDKITE_GHA_MISE", filepath.Join(t.TempDir(), "missing-mise"))
-	runner := &cliCaptureRunner{}
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "prepare action runtime: resolve runtime mise executable") {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if !strings.Contains(stdout.String(), "+++ :warning: Prepare GitHub Actions job failed\n~~~ :package: Publish GitHub Actions result\n") {
-		t.Fatalf("stdout = %q, want visible runner setup failure before collapsed publication", stdout.String())
-	}
-	if manifest := publishedCLIManifest(t, runner, job, planDigest); manifest.Result != "failure" {
-		t.Fatalf("published result = %q, want failure", manifest.Result)
-	}
-}
-
-func TestRunJobMiseRequirementUsesCompilerDecisionAndFailsClosed(t *testing.T) {
-	no := false
-	yes := true
-	actionJob := func(requiresMise *bool) plan.Job {
-		return plan.Job{
-			Actions:      []plan.ActionLock{{ID: "a-0000000000000001"}},
-			Steps:        []plan.Step{{Kind: "uses", Action: &plan.ActionSelector{Lock: "a-0000000000000001"}}},
-			RequiresMise: requiresMise,
-		}
-	}
-	cacheJob := actionJob(&yes)
-	cacheJob.Actions[0].Source = "github"
-	cacheJob.Actions[0].Repository = "actions/cache"
-	for _, test := range []struct {
-		name string
-		job  plan.Job
-		want bool
-	}{
-		{name: "shell plan does not resolve mise", job: plan.Job{Steps: []plan.Step{{Kind: "run"}}}},
-		{name: "native plan does not resolve mise", job: actionJob(&no)},
-		{name: "JavaScript plan resolves mise", job: actionJob(&yes), want: true},
-		{name: "cache client plan resolves mise", job: cacheJob, want: true},
-		{name: "missing action decision resolves mise", job: actionJob(nil), want: true},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			if got := test.job.NeedsMise(); got != test.want {
-				t.Fatalf("NeedsMise() = %t, want %t", got, test.want)
-			}
-		})
-	}
-}
-
-func TestRunJobCallGuardSkipsActionJobBeforePreparingRuntimeMise(t *testing.T) {
-	job := cliRunJobPlan()
-	job.Schema = plan.Schema
-	job.Condition = "always()"
-	job.CallGuards = []plan.CallGuard{{Condition: "false"}}
-	job.Actions = []plan.ActionLock{{
-		ID: "a-0000000000000001", Source: "workspace", Path: "actions/build",
-		SourceDigest: "sha256:" + strings.Repeat("a", 64),
-	}}
-	job.Steps = []plan.Step{{
-		ID: "local", Kind: "uses", Uses: "./actions/build",
-		Action: &plan.ActionSelector{Lock: "a-0000000000000001"},
-	}}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	t.Setenv("BUILDKITE_GHA_MISE", filepath.Join(t.TempDir(), "missing-mise"))
-	runner := &cliCaptureRunner{}
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 0 {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if strings.Contains(stderr.String(), "prepare action runtime") {
-		t.Fatalf("skipped job prepared action runtime: %q", stderr.String())
-	}
-	if manifest := publishedCLIManifest(t, runner, job, planDigest); manifest.Result != "skipped" {
-		t.Fatalf("published result = %q, want skipped", manifest.Result)
-	}
 }
 
 func TestRunUploadFailsClosedBeforePipeline(t *testing.T) {
@@ -7233,633 +6724,6 @@ func (r *cliCaptureRunner) Run(ctx context.Context, dir, name string, args []str
 
 var _ transport.Runner = (*cliCaptureRunner)(nil)
 
-func TestRunJobExecutesBoundPlanAndWritesResult(t *testing.T) {
-	workspace := t.TempDir()
-	workflowSource := []byte("name: cli fixture\n")
-	workflowPath := filepath.Join(workspace, "workflow.yml")
-	if err := os.WriteFile(workflowPath, workflowSource, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	digest := sha256.Sum256(workflowSource)
-	requiresMise := false
-	job := plan.Job{
-		Schema: plan.Schema, Compiler: plan.Compiler{Version: "dev", DistributionDigest: "sha256:" + strings.Repeat("9", 64)},
-		Runtime:      &plan.Runtime{DistributionDigest: cliTestRuntimeDigest()},
-		Workflow:     plan.Workflow{Path: "workflow.yml", Digest: "sha256:" + hex.EncodeToString(digest[:]), LogicalJobID: "cli"},
-		Event:        plan.Event{Provider: "github", Name: "push", PayloadDigest: "sha256:" + strings.Repeat("3", 64)},
-		Target:       plan.Target{StepKey: "gha-cli", Queue: "ubuntu-latest"},
-		Outputs:      map[string]string{"result": "${{ steps.produce.outputs.result }}"},
-		Steps:        []plan.Step{{ID: "produce", Kind: "run", Command: `echo "result=cli-ok" >> "$GITHUB_OUTPUT"`}},
-		RequiresMise: &requiresMise,
-	}
-	encoded, err := plan.Encode(job)
-	if err != nil {
-		t.Fatal(err)
-	}
-	planPath := filepath.Join(workspace, "plan.json")
-	resultPath := filepath.Join(workspace, "result.json")
-	if err := os.WriteFile(planPath, encoded, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	planDigest := sha256.Sum256(encoded)
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "sha256:"+hex.EncodeToString(planDigest[:]))
-	t.Setenv("BUILDKITE", "true")
-	t.Setenv("BUILDKITE_BUILD_ID", cliTestBuildID)
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_STEP_KEY", job.Target.StepKey)
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", job.Target.Queue)
-	oldDirectory, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Chdir(workspace); err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { _ = os.Chdir(oldDirectory) })
-
-	var stdout, stderr bytes.Buffer
-	runner := &cliCaptureRunner{}
-	if code := run([]string{"run-job", "--plan", planPath, "--result", resultPath}, &stdout, &stderr, "dev", runner); code != 0 {
-		t.Fatalf("Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	result, err := os.ReadFile(resultPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !strings.Contains(string(result), `"result": "cli-ok"`) {
-		t.Fatalf("result = %s", result)
-	}
-	artifactDigest := transport.Digest(encoded)
-	artifactPath, err := buildkitepipeline.PlanPath(artifactDigest)
-	if err != nil {
-		t.Fatal(err)
-	}
-	artifactRunner := &cliCaptureRunner{dataByPath: map[string][]byte{artifactPath: encoded}}
-	artifactResultPath := filepath.Join(workspace, "artifact-result.json")
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "sha256:"+strings.Repeat("0", 64))
-	stdout.Reset()
-	stderr.Reset()
-	failedArtifactRunner := &cliCaptureRunner{failAt: 1}
-	if code := run([]string{"run-job", "--plan-digest", artifactDigest, "--plan-producer", "gha-importer", "--result", artifactResultPath}, &stdout, &stderr, "dev", failedArtifactRunner); code != 1 || !strings.Contains(stderr.String(), "download plan") {
-		t.Fatalf("failed artifact Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if !strings.Contains(stdout.String(), "~~~ :package: Prepare GitHub Actions job\n^^^ +++\n") {
-		t.Fatalf("failed artifact stdout = %q, want expanded preparation group", stdout.String())
-	}
-	stdout.Reset()
-	stderr.Reset()
-	if code := run([]string{"run-job", "--plan-digest", artifactDigest, "--plan-producer", "gha-importer", "--result", artifactResultPath}, &stdout, &stderr, "dev", artifactRunner); code != 0 {
-		t.Fatalf("artifact Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if len(artifactRunner.commands) == 0 || !slices.Equal(artifactRunner.commands[0].args[:3], []string{"artifact", "download", artifactPath}) || artifactRunner.commands[0].args[4] != "--step" || artifactRunner.commands[0].args[5] != "gha-importer" {
-		t.Fatalf("plan artifact download = %#v", artifactRunner.commands)
-	}
-	if destination := artifactRunner.commands[0].args[3]; destination == workspace {
-		t.Fatalf("plan downloaded into workspace %q", destination)
-	} else if _, err := os.Stat(destination); !os.IsNotExist(err) {
-		t.Fatalf("private plan directory still exists: %v", err)
-	}
-	tamperedRunner := &cliCaptureRunner{dataByPath: map[string][]byte{artifactPath: []byte("tampered")}}
-	stdout.Reset()
-	stderr.Reset()
-	if code := run([]string{"run-job", "--plan-digest", artifactDigest, "--plan-producer", "gha-importer"}, &stdout, &stderr, "dev", tamperedRunner); code != 1 || !strings.Contains(stderr.String(), "does not match expected digest") {
-		t.Fatalf("tampered artifact Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	stdout.Reset()
-	stderr.Reset()
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "sha256:"+strings.Repeat("0", 64))
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "does not match expected digest") {
-		t.Fatalf("Run() code = %d, stderr = %q, want digest mismatch", code, stderr.String())
-	}
-	job.Runtime.DistributionDigest = "sha256:" + strings.Repeat("0", 64)
-	encoded, err = plan.Encode(job)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(planPath, encoded, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	planDigest = sha256.Sum256(encoded)
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "sha256:"+hex.EncodeToString(planDigest[:]))
-	stdout.Reset()
-	stderr.Reset()
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "runtime distribution digest") {
-		t.Fatalf("Run() code = %d, stderr = %q, want runtime digest mismatch", code, stderr.String())
-	}
-	job.Runtime.DistributionDigest = cliTestRuntimeDigest()
-	job.Compiler.Version = "0.0.0-other"
-	encoded, err = plan.Encode(job)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(planPath, encoded, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	planDigest = sha256.Sum256(encoded)
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "sha256:"+hex.EncodeToString(planDigest[:]))
-	stdout.Reset()
-	stderr.Reset()
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "does not match runtime version") {
-		t.Fatalf("Run() code = %d, stderr = %q, want version mismatch", code, stderr.String())
-	}
-}
-
-func TestRunJobHydratesNeedsAndPublishesAuthoritativeResult(t *testing.T) {
-	producerStep := "gha-producer"
-	producerPlanDigest := transport.Digest([]byte("producer-plan"))
-	job := cliRunJobPlan()
-	job.Dependencies = []string{producerStep}
-	job.NeedSources = map[string][]plan.NeedSource{
-		"producer": {{StepKey: producerStep, PlanDigest: producerPlanDigest}},
-	}
-	job.CallGuards = []plan.CallGuard{{
-		Condition:   "needs.producer.result == 'success' && needs.producer.outputs.result == 'hydrated'",
-		NeedSources: map[string][]plan.NeedSource{"producer": {{StepKey: producerStep, PlanDigest: producerPlanDigest}}},
-		NeedOutputs: map[string][]plan.NeedOutput{"producer": {{Name: "result", StepKey: producerStep, Output: "result"}}},
-	}}
-	job.Env = map[string]string{"RESULT": "${{ needs.producer.outputs.result }}"}
-	job.Steps = []plan.Step{{ID: "consume", Kind: "run", Command: `test "$RESULT" = "hydrated"`}}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-
-	producerManifest, err := transport.MarshalResultManifest(transport.ResultManifest{
-		PlanDigest: producerPlanDigest,
-		Producer:   transport.Producer{BuildID: cliTestBuildID, JobID: cliTestProducerJobID, StepKey: producerStep},
-		Result:     "success",
-		Outputs:    []transport.Output{{Name: "result", Value: "hydrated"}},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	producerPath := transport.ResultPath(producerStep, producerPlanDigest)
-	runner := &cliCaptureRunner{
-		failMetadata: true,
-		jobByStep:    map[string]string{producerStep: cliTestProducerJobID},
-		dataByPath:   map[string][]byte{producerPath: producerManifest},
-	}
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 0 {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if !strings.Contains(stderr.String(), "warning: result metadata mirror") {
-		t.Fatalf("stderr = %q, want non-fatal metadata warning", stderr.String())
-	}
-	if !strings.Contains(stdout.String(), "~~~ :package: Publish GitHub Actions result\n") || !strings.Contains(stdout.String(), "^^^ +++\n") {
-		t.Fatalf("stdout = %q, want expanded result publication group", stdout.String())
-	}
-	if len(runner.commands) < 3 || strings.Join(runner.commands[0].args, " ") != strings.Join([]string{"artifact", "search", producerPath, "--step", producerStep, "--format", "%j"}, " ") {
-		t.Fatalf("commands = %#v, want exact producer search first", runner.commands)
-	}
-	if got := runner.commands[1].args[len(runner.commands[1].args)-1]; got != cliTestProducerJobID {
-		t.Fatalf("download producer = %q, want exact job UUID", got)
-	}
-	manifest := publishedCLIManifest(t, runner, job, planDigest)
-	if manifest.Result != "success" {
-		t.Fatalf("published result = %q, want success", manifest.Result)
-	}
-	for _, command := range runner.commands {
-		if command.dir != "" {
-			if _, err := os.Stat(command.dir); !os.IsNotExist(err) {
-				t.Fatalf("temporary artifact root %q still exists: %v", command.dir, err)
-			}
-		}
-	}
-}
-
-func TestRunJobPublishesSummaryAsAdvisoryJobAnnotation(t *testing.T) {
-	tests := []struct {
-		name           string
-		command        string
-		failAnnotation bool
-		wantAnnotation bool
-		wantWarning    bool
-		wantResult     string
-		wantCode       int
-	}{
-		{name: "published", command: `printf '### Job summary\n\nPassed.\n' >> "$GITHUB_STEP_SUMMARY"`, wantAnnotation: true, wantResult: "success"},
-		{name: "published after job failure", command: `printf '### Job summary\n\nPassed.\n' >> "$GITHUB_STEP_SUMMARY"; exit 7`, wantAnnotation: true, wantResult: "failure", wantCode: 1},
-		{name: "failure is advisory", command: `printf '### Job summary\n\nPassed.\n' >> "$GITHUB_STEP_SUMMARY"`, failAnnotation: true, wantAnnotation: true, wantWarning: true, wantResult: "success"},
-		{name: "empty summary is a no-op", command: "true", wantResult: "success"},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			job := cliRunJobPlan()
-			job.Steps[0].Command = test.command
-			planPath, planDigest := writeCLIJobPlan(t, job)
-			setCLIJobIdentity(t, job, planDigest)
-			runner := &cliCaptureRunner{failAnnotation: test.failAnnotation}
-			var stdout, stderr bytes.Buffer
-
-			if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != test.wantCode {
-				t.Fatalf("run() code = %d, stderr = %q, want %d", code, stderr.String(), test.wantCode)
-			}
-			if !strings.Contains(stdout.String(), "~~~ :package: Publish GitHub Actions result\n") {
-				t.Fatalf("stdout = %q, want result publication group", stdout.String())
-			}
-			if test.wantWarning && !strings.Contains(stdout.String(), "^^^ +++\n") {
-				t.Fatalf("stdout = %q, want publication warning to expand its group", stdout.String())
-			}
-			if !test.wantWarning && test.wantCode == 0 && strings.Contains(stdout.String(), "^^^ +++\n") {
-				t.Fatalf("stdout = %q, want successful publication group collapsed", stdout.String())
-			}
-			if test.wantCode != 0 && strings.Contains(stdout.String(), "Prepare GitHub Actions job failed") {
-				t.Fatalf("stdout = %q, want action failure owned by its existing section", stdout.String())
-			}
-			if result := publishedCLIManifest(t, runner, job, planDigest); result.Result != test.wantResult {
-				t.Fatalf("published result = %q, want %q", result.Result, test.wantResult)
-			}
-			var annotations []cliCommand
-			for _, command := range runner.commands {
-				if len(command.args) > 0 && command.args[0] == "annotate" {
-					annotations = append(annotations, command)
-				}
-			}
-			if !test.wantAnnotation {
-				if len(annotations) != 0 {
-					t.Fatalf("annotations = %#v, want none", annotations)
-				}
-				return
-			}
-			wantArgs := []string{"annotate", "--scope", "job", "--job", cliTestJobID, "--context", "buildkite-gha-job-summary", "--style", "info"}
-			wantBody := "<h2 class=\"h4 mb2\">GitHub Actions job summary</h2>\n<div class=\"border-top border-gray pt2\"></div>\n\n### Job summary\n\nPassed.\n"
-			if len(annotations) != 1 || !slices.Equal(annotations[0].args, wantArgs) || string(annotations[0].stdin) != wantBody {
-				t.Fatalf("annotations = %#v", annotations)
-			}
-			if last := runner.commands[len(runner.commands)-1]; len(last.args) == 0 || last.args[0] != "annotate" {
-				t.Fatalf("last command = %#v, want annotation after authoritative publication", last)
-			}
-			if gotWarning := strings.Contains(stderr.String(), "warning: job summary annotation"); gotWarning != test.wantWarning {
-				t.Fatalf("stderr = %q, warning present = %v, want %v", stderr.String(), gotWarning, test.wantWarning)
-			}
-			if runner.contextErrors[len(runner.contextErrors)-1] != nil {
-				t.Fatalf("annotation inherited cancelled context: %v", runner.contextErrors[len(runner.contextErrors)-1])
-			}
-		})
-	}
-}
-
-func TestRunJobDisabledWorkflowTokenLinksPipelineSettings(t *testing.T) {
-	server := httptest.NewServer(http.NotFoundHandler())
-	defer server.Close()
-
-	job := cliRunJobPlan()
-	job.Workflow.Path = ".github/workflows/ci.yml"
-	job.Event.Repository = "buildkite/buildkite-gha"
-	job.RequiredCapabilities = []string{"provider-token-write"}
-	job.GitHubToken = &plan.GitHubToken{Workflow: "ci.yml", Permissions: map[string]string{"contents": "read"}}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	truePath, err := exec.LookPath("true")
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "job-token")
-	t.Setenv("BUILDKITE_GHA_AGENT", truePath)
-	t.Setenv("BUILDKITE_ORGANIZATION_SLUG", "acme-inc")
-	t.Setenv("BUILDKITE_PIPELINE_SLUG", "my-pipeline")
-
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", &cliCaptureRunner{}); code != 1 {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	want := `enable "Allow workflow-authorized GitHub access tokens" in the pipeline's repository settings: https://buildkite.com/acme-inc/my-pipeline/settings/repository`
-	if !strings.Contains(stderr.String(), want) {
-		t.Fatalf("stderr = %q, want %q", stderr.String(), want)
-	}
-}
-
-func TestRunJobPublishesWorkflowCommandsAsAdvisoryJobAnnotations(t *testing.T) {
-	diagnostics := "printf '%s\\n' '::warning title=Lint::warning body'; printf '%s\\n' '::error file=main.go,line=7::error body' >&2"
-	tests := []struct {
-		name           string
-		command        string
-		failAnnotation bool
-		wantAnnotation bool
-		wantWarning    bool
-		wantResult     string
-		wantCode       int
-	}{
-		{name: "published without changing success", command: diagnostics, wantAnnotation: true, wantResult: "success"},
-		{name: "published after job failure", command: diagnostics + "; exit 7", wantAnnotation: true, wantResult: "failure", wantCode: 1},
-		{name: "publication failure is advisory", command: diagnostics, failAnnotation: true, wantAnnotation: true, wantWarning: true, wantResult: "success"},
-		{name: "empty diagnostics are a no-op", command: "true", wantResult: "success"},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			job := cliRunJobPlan()
-			job.Steps[0].Command = test.command
-			planPath, planDigest := writeCLIJobPlan(t, job)
-			setCLIJobIdentity(t, job, planDigest)
-			runner := &cliCaptureRunner{failAnnotation: test.failAnnotation}
-			var stdout, stderr bytes.Buffer
-
-			if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != test.wantCode {
-				t.Fatalf("run() code = %d, stderr = %q, want %d", code, stderr.String(), test.wantCode)
-			}
-			if result := publishedCLIManifest(t, runner, job, planDigest); result.Result != test.wantResult {
-				t.Fatalf("published result = %q, want %q", result.Result, test.wantResult)
-			}
-			var annotations []cliCommand
-			for _, command := range runner.commands {
-				if len(command.args) > 0 && command.args[0] == "annotate" {
-					annotations = append(annotations, command)
-				}
-			}
-			if !test.wantAnnotation {
-				if len(annotations) != 0 {
-					t.Fatalf("annotations = %#v, want none", annotations)
-				}
-				return
-			}
-			want := []struct {
-				context, style, body string
-			}{
-				{context: "buildkite-gha-workflow-warnings", style: "warning", body: "warning body"},
-				{context: "buildkite-gha-workflow-errors", style: "error", body: "error body"},
-			}
-			if len(annotations) != len(want) {
-				t.Fatalf("annotations = %#v, want warning and error", annotations)
-			}
-			for i, expected := range want {
-				wantArgs := []string{"annotate", "--scope", "job", "--job", cliTestJobID, "--context", expected.context, "--style", expected.style}
-				if !slices.Equal(annotations[i].args, wantArgs) || !strings.Contains(string(annotations[i].stdin), expected.body) {
-					t.Errorf("annotation %d = %#v, want args %#v and body containing %q", i, annotations[i], wantArgs, expected.body)
-				}
-			}
-			for _, command := range runner.commands[len(runner.commands)-len(want):] {
-				if len(command.args) == 0 || command.args[0] != "annotate" {
-					t.Fatalf("trailing command = %#v, want annotations after authoritative publication", command)
-				}
-			}
-			gotWarning := strings.Contains(stderr.String(), "warning: workflow warning annotation") && strings.Contains(stderr.String(), "warning: workflow error annotation")
-			if gotWarning != test.wantWarning {
-				t.Fatalf("stderr = %q, advisory warnings present = %v, want %v", stderr.String(), gotWarning, test.wantWarning)
-			}
-			for _, contextErr := range runner.contextErrors[len(runner.contextErrors)-len(want):] {
-				if contextErr != nil {
-					t.Fatalf("annotation inherited cancelled context: %v", contextErr)
-				}
-			}
-		})
-	}
-}
-
-func TestRunJobPublishesEveryTerminalResultAfterCancellation(t *testing.T) {
-	tests := []struct {
-		name       string
-		condition  string
-		command    string
-		cancel     bool
-		tolerate   bool
-		wantResult string
-		wantCode   int
-	}{
-		{name: "failure", command: "exit 7", wantResult: "failure", wantCode: 1},
-		{name: "tolerated failure", command: "exit 7", tolerate: true, wantResult: "success", wantCode: buildkitepipeline.ContinueOnErrorExitStatus},
-		{name: "skipped", condition: "${{ false }}", command: "true", wantResult: "skipped", wantCode: 0},
-		{name: "cancelled", command: "sleep 1", cancel: true, wantResult: "cancelled", wantCode: 1},
-		{name: "tolerated cancellation", command: "sleep 1", cancel: true, tolerate: true, wantResult: "cancelled", wantCode: 1},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			job := cliRunJobPlan()
-			job.Condition = test.condition
-			job.Steps[0].Command = test.command
-			if test.tolerate {
-				job.Schema = plan.Schema
-				job.ContinueOnError = true
-				job.Runtime = &plan.Runtime{DistributionDigest: cliTestRuntimeDigest()}
-				requiresMise := false
-				job.RequiresMise = &requiresMise
-			}
-			planPath, planDigest := writeCLIJobPlan(t, job)
-			setCLIJobIdentity(t, job, planDigest)
-			runner := &cliCaptureRunner{}
-			ctx := context.Background()
-			if test.cancel {
-				cancelled, cancel := context.WithCancel(ctx)
-				cancel()
-				ctx = cancelled
-			}
-			var stdout, stderr bytes.Buffer
-			if code := runJobContext(ctx, []string{"--plan", planPath}, &stdout, &stderr, "dev", "dev", transport.Agent{Runner: runner}); code != test.wantCode {
-				t.Fatalf("runJobContext() code = %d, stderr = %q, want %d", code, stderr.String(), test.wantCode)
-			}
-			manifest := publishedCLIManifest(t, runner, job, planDigest)
-			if manifest.Result != test.wantResult {
-				t.Fatalf("published result = %q, want %q", manifest.Result, test.wantResult)
-			}
-			for i, contextErr := range runner.contextErrors {
-				if contextErr != nil {
-					t.Fatalf("publication command %d inherited cancelled context: %v", i, contextErr)
-				}
-			}
-		})
-	}
-}
-
-func TestPublishTerminalResultAnnotatesCancelledJobWithFreshContext(t *testing.T) {
-	job := cliRunJobPlan()
-	planDigest := transport.Digest([]byte("cancelled-plan"))
-	producer := transport.Producer{BuildID: cliTestBuildID, JobID: cliTestJobID, StepKey: job.Target.StepKey}
-	runner := &cliCaptureRunner{}
-	artifactDigest := strings.Repeat("a", 64)
-	artifact := transport.ResultArtifact{
-		Name: "payload", ID: "123", Path: "buildkite-gha/v1/artifacts/" + artifactDigest + ".zip",
-		Digest: "sha256:" + artifactDigest, Size: 42, FileCount: 1,
-	}
-
-	publication, err := publishTerminalResult(
-		transport.Agent{Runner: runner},
-		t.TempDir(),
-		job,
-		planDigest,
-		producer,
-		gharuntime.JobResult{
-			Conclusion:         "cancelled",
-			Summary:            "summary before cancellation\n",
-			WarningAnnotations: "warning before cancellation\n",
-			ErrorAnnotations:   "error before cancellation\n",
-			Artifacts:          []transport.ResultArtifact{artifact},
-		},
-	)
-	if err != nil || publication.SummaryAnnotationError != nil || publication.WarningAnnotationError != nil || publication.ErrorAnnotationError != nil {
-		t.Fatalf("publishTerminalResult() publication = %#v, error = %v", publication, err)
-	}
-	wantBodies := []string{
-		"<h2 class=\"h4 mb2\">GitHub Actions job summary</h2>\n<div class=\"border-top border-gray pt2\"></div>\n\nsummary before cancellation\n",
-		"warning before cancellation\n",
-		"error before cancellation\n",
-	}
-	commands := runner.commands[len(runner.commands)-len(wantBodies):]
-	for i, command := range commands {
-		if len(command.args) == 0 || command.args[0] != "annotate" || string(command.stdin) != wantBodies[i] {
-			t.Errorf("annotation %d = %#v, want cancelled job annotation body %q", i, command, wantBodies[i])
-		}
-	}
-	for _, contextErr := range runner.contextErrors[len(runner.contextErrors)-len(wantBodies):] {
-		if contextErr != nil {
-			t.Fatalf("annotation inherited cancelled context: %v", contextErr)
-		}
-	}
-	manifest := publishedCLIManifest(t, runner, job, planDigest)
-	if !reflect.DeepEqual(manifest.Artifacts, []transport.ResultArtifact{artifact}) {
-		t.Fatalf("published artifacts = %#v, want %#v", manifest.Artifacts, artifact)
-	}
-}
-
-func TestRunJobPublishesHydrationFailureAndRejectsMissingIdentity(t *testing.T) {
-	producerStep := "gha-producer"
-	producerPlanDigest := transport.Digest([]byte("producer-plan"))
-	job := cliRunJobPlan()
-	job.Dependencies = []string{producerStep}
-	job.NeedSources = map[string][]plan.NeedSource{
-		"producer": {{StepKey: producerStep, PlanDigest: producerPlanDigest}},
-	}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-
-	t.Run("needs require Buildkite", func(t *testing.T) {
-		clearCLIJobIdentity(t)
-		runner := &cliCaptureRunner{}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "prerequisites require Buildkite result identity") {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if len(runner.commands) != 0 {
-			t.Fatalf("commands = %#v, want no unauthenticated agent calls", runner.commands)
-		}
-	})
-
-	t.Run("hydration failure is published", func(t *testing.T) {
-		setCLIJobIdentity(t, job, planDigest)
-		runner := &cliCaptureRunner{jobByStep: map[string]string{producerStep: cliTestProducerJobID}}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "hydrate prerequisite results") {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if !strings.Contains(stdout.String(), "+++ :warning: Prepare GitHub Actions job failed\n~~~ :package: Publish GitHub Actions result\n") {
-			t.Fatalf("stdout = %q, want visible prerequisite failure before collapsed publication", stdout.String())
-		}
-		manifest := publishedCLIManifest(t, runner, job, planDigest)
-		if manifest.Result != "failure" {
-			t.Fatalf("published result = %q, want failure", manifest.Result)
-		}
-	})
-
-	t.Run("partial identity fails before execution", func(t *testing.T) {
-		setCLIJobIdentity(t, job, planDigest)
-		t.Setenv("BUILDKITE_JOB_ID", "")
-		runner := &cliCaptureRunner{}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "requires valid BUILDKITE_BUILD_ID") {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if len(runner.commands) != 0 {
-			t.Fatalf("commands = %#v, want identity failure before side effects", runner.commands)
-		}
-	})
-
-	t.Run("missing plan digest fails before execution", func(t *testing.T) {
-		setCLIJobIdentity(t, job, planDigest)
-		t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", "")
-		runner := &cliCaptureRunner{}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "requires BUILDKITE_GHA_PLAN_DIGEST") {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if len(runner.commands) != 0 {
-			t.Fatalf("commands = %#v, want digest failure before side effects", runner.commands)
-		}
-	})
-}
-
-func TestRunJobFailsWhenAuthoritativePublicationFails(t *testing.T) {
-	job := cliRunJobPlan()
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	events := captureCommandTelemetry(t)
-	runner := &cliCaptureRunner{failAt: 1}
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "publish terminal result") {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if !strings.Contains(stdout.String(), "~~~ :package: Publish GitHub Actions result\n") || !strings.Contains(stdout.String(), "^^^ +++\n") {
-		t.Fatalf("stdout = %q, want failed publication group expanded", stdout.String())
-	}
-	if event := <-events; event.FailurePhase != telemetry.FailurePhaseResultPublication || event.FailureCode != telemetry.FailureCodeUnknown {
-		t.Fatalf("telemetry = %#v", event)
-	}
-}
-
-func TestRunJobTelemetryClassifiesExecutionFailure(t *testing.T) {
-	job := cliRunJobPlan()
-	job.Steps[0].Command = "exit 7"
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	events := captureCommandTelemetry(t)
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", &cliCaptureRunner{}); code != 1 {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	if event := <-events; event.FailurePhase != telemetry.FailurePhaseExecution || event.FailureCode != telemetry.FailureCodeUnknown {
-		t.Fatalf("telemetry = %#v", event)
-	}
-}
-
-func TestRunJobPublishesBoundedFailureForUnrepresentableOutputs(t *testing.T) {
-	job := cliRunJobPlan()
-	job.Steps[0].Command = `printf 'summary from malformed result\n' >> "$GITHUB_STEP_SUMMARY"`
-	job.Outputs = make(map[string]string, transport.MaxResultOutputs+1)
-	for i := 0; i <= transport.MaxResultOutputs; i++ {
-		job.Outputs[fmt.Sprintf("output_%02d", i)] = "value"
-	}
-	planPath, planDigest := writeCLIJobPlan(t, job)
-	setCLIJobIdentity(t, job, planDigest)
-	runner := &cliCaptureRunner{}
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--plan", planPath}, &stdout, &stderr, "dev", runner); code != 1 || !strings.Contains(stderr.String(), "validate terminal result") {
-		t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-	}
-	manifest := publishedCLIManifest(t, runner, job, planDigest)
-	if manifest.Result != "failure" || len(manifest.Outputs) != 0 {
-		t.Fatalf("published result = %#v, want bounded failure without outputs", manifest)
-	}
-	last := runner.commands[len(runner.commands)-1]
-	wantSummary := "<h2 class=\"h4 mb2\">GitHub Actions job summary</h2>\n<div class=\"border-top border-gray pt2\"></div>\n\nsummary from malformed result\n"
-	if len(last.args) == 0 || last.args[0] != "annotate" || string(last.stdin) != wantSummary {
-		t.Fatalf("last command = %#v, want summary annotation after bounded failure", last)
-	}
-}
-
-func TestVerifyBuildkiteTargetFailsClosed(t *testing.T) {
-	job := plan.Job{Target: plan.Target{StepKey: "gha-expected", Queue: "gha-runtime"}}
-	t.Setenv("BUILDKITE", "true")
-	t.Setenv("BUILDKITE_STEP_KEY", "gha-other")
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", "gha-runtime")
-	if err := verifyBuildkiteTarget(job); err == nil || !strings.Contains(err.Error(), "executing step") {
-		t.Fatalf("verifyBuildkiteTarget() error = %v, want step mismatch", err)
-	}
-	t.Setenv("BUILDKITE_STEP_KEY", "")
-	if err := verifyBuildkiteTarget(job); err == nil || !strings.Contains(err.Error(), "requires BUILDKITE_STEP_KEY") {
-		t.Fatalf("verifyBuildkiteTarget() error = %v, want missing binding", err)
-	}
-	t.Setenv("BUILDKITE_STEP_KEY", "gha-expected")
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", "gha-other")
-	if err := verifyBuildkiteTarget(job); err == nil || !strings.Contains(err.Error(), "executing queue") {
-		t.Fatalf("verifyBuildkiteTarget() error = %v, want explicit queue mismatch", err)
-	}
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", "")
-	if err := verifyBuildkiteTarget(job); err == nil || !strings.Contains(err.Error(), "requires BUILDKITE_AGENT_META_DATA_QUEUE") {
-		t.Fatalf("verifyBuildkiteTarget() error = %v, want missing explicit queue binding", err)
-	}
-	job.Target.Queue = ""
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", "customer-default")
-	if err := verifyBuildkiteTarget(job); err != nil {
-		t.Fatalf("verifyBuildkiteTarget() default targeting error = %v", err)
-	}
-}
-
 func TestArgumentParsersRejectRepeatedOptions(t *testing.T) {
 	if _, _, err := validateActionCacheArgs([]string{"--action-cache-dir", "one", "--action-cache-dir", "two", "workflow.yml"}); err == nil || !strings.Contains(err.Error(), "only be specified once") {
 		t.Fatalf("validateActionCacheArgs() error = %v, want duplicate cache directory error", err)
@@ -8072,175 +6936,10 @@ func TestLoadRuntimeDistributionsValidatesPlatformBinaryAndSymlink(t *testing.T)
 	}
 }
 
-func TestRepositoryProviderGitCredentialsUseServerEnvironment(t *testing.T) {
-	values := map[string]string{}
-	getenv := func(name string) string { return values[name] }
-	if repositoryProviderGitCredentialsEnabled(getenv) {
-		t.Fatal("repository provider credentials enabled without a server setting")
-	}
-	values[repositoryProviderGitCredentialsEnvironment] = "true"
-	if !repositoryProviderGitCredentialsEnabled(getenv) {
-		t.Fatal("repository provider credentials setting was ignored")
-	}
-	delete(values, repositoryProviderGitCredentialsEnvironment)
-	values[legacyGitHubAppGitCredentialsEnvironment] = "true"
-	if !repositoryProviderGitCredentialsEnabled(getenv) {
-		t.Fatal("legacy GitHub App credentials setting was ignored")
-	}
-	values[legacyGitHubAppGitCredentialsEnvironment] = "TRUE"
-	if repositoryProviderGitCredentialsEnabled(getenv) {
-		t.Fatal("non-canonical server setting enabled repository credentials")
-	}
-}
-
-func TestRunJobExecutesPureRunPlanWithoutCheckout(t *testing.T) {
-	clearCLIJobIdentity(t)
-	workspace := t.TempDir()
-	requiresMise := false
-	job := plan.Job{
-		Schema: plan.Schema,
-		Compiler: plan.Compiler{
-			Version:            "dev",
-			DistributionDigest: cliTestRuntimeDigest(),
-		},
-		Runtime: &plan.Runtime{DistributionDigest: cliTestRuntimeDigest()},
-		Workflow: plan.Workflow{
-			Path:         "missing-workflow.yml",
-			Digest:       "sha256:" + strings.Repeat("1", 64),
-			LogicalJobID: "missing",
-		},
-		Event:                plan.Event{Provider: "github", Name: "push", PayloadDigest: "sha256:" + strings.Repeat("3", 64)},
-		Target:               plan.Target{StepKey: "gha-missing", Queue: "gha-runtime"},
-		RequiredCapabilities: []string{},
-		Steps:                []plan.Step{{ID: "step-1", Kind: "run", Command: "true"}},
-		RequiresMise:         &requiresMise,
-	}
-	encoded, err := plan.Encode(job)
-	if err != nil {
-		t.Fatal(err)
-	}
-	planPath := filepath.Join(workspace, "plan.json")
-	resultPath := filepath.Join(workspace, "result.json")
-	if err := os.WriteFile(planPath, encoded, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	oldDirectory, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Chdir(workspace); err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { _ = os.Chdir(oldDirectory) })
-
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"run-job", "--plan", planPath, "--result", resultPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, stderr = %q, want success", code, stderr.String())
-	}
-	result, err := os.ReadFile(resultPath)
-	if err != nil || !bytes.Contains(result, []byte(`"conclusion": "success"`)) {
-		t.Fatalf("result = %q, error = %v", result, err)
-	}
-}
-
-func cliRunJobPlan() plan.Job {
-	requiresMise := false
-	return plan.Job{
-		Schema: plan.Schema,
-		Compiler: plan.Compiler{
-			Version:            "dev",
-			DistributionDigest: cliTestRuntimeDigest(),
-		},
-		Runtime: &plan.Runtime{DistributionDigest: cliTestRuntimeDigest()},
-		Workflow: plan.Workflow{
-			Path:         "workflow.yml",
-			Digest:       "sha256:" + strings.Repeat("1", 64),
-			LogicalJobID: "cli",
-		},
-		Event:                plan.Event{Provider: "github", Name: "push", PayloadDigest: "sha256:" + strings.Repeat("3", 64)},
-		Target:               plan.Target{StepKey: "gha-cli", Queue: "gha-runtime"},
-		RequiredCapabilities: []string{},
-		Steps:                []plan.Step{{ID: "step-1", Kind: "run", Command: "true"}},
-		RequiresMise:         &requiresMise,
-	}
-}
-
 func cliTestRuntimeDigest() string {
 	digest, err := executableDigest()
 	if err != nil {
 		panic(err)
 	}
 	return digest
-}
-
-func writeCLIJobPlan(t *testing.T, job plan.Job) (string, string) {
-	t.Helper()
-	encoded, err := plan.Encode(job)
-	if err != nil {
-		t.Fatal(err)
-	}
-	path := filepath.Join(t.TempDir(), "plan.json")
-	if err := os.WriteFile(path, encoded, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	return path, transport.Digest(encoded)
-}
-
-func setCLIJobIdentity(t *testing.T, job plan.Job, planDigest string) {
-	t.Helper()
-	t.Setenv("BUILDKITE", "true")
-	t.Setenv("BUILDKITE_BUILD_ID", cliTestBuildID)
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_STEP_KEY", job.Target.StepKey)
-	t.Setenv("BUILDKITE_AGENT_META_DATA_QUEUE", job.Target.Queue)
-	t.Setenv("BUILDKITE_GHA_PLAN_DIGEST", planDigest)
-}
-
-func captureCommandTelemetry(t *testing.T) <-chan telemetry.Properties {
-	t.Helper()
-	events := make(chan telemetry.Properties, 1)
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var received struct {
-			Properties telemetry.Properties `json:"properties"`
-		}
-		if err := json.NewDecoder(r.Body).Decode(&received); err != nil {
-			t.Errorf("decode telemetry event: %v", err)
-		}
-		events <- received.Properties
-		w.WriteHeader(http.StatusAccepted)
-	}))
-	t.Cleanup(server.Close)
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "telemetry-token")
-	t.Setenv("BUILDKITE_GHA_TELEMETRY_DISABLED", "")
-	return events
-}
-
-func clearCLIJobIdentity(t *testing.T) {
-	t.Helper()
-	for _, name := range []string{
-		"BUILDKITE",
-		"BUILDKITE_BUILD_ID",
-		"BUILDKITE_JOB_ID",
-		"BUILDKITE_STEP_KEY",
-		"BUILDKITE_AGENT_META_DATA_QUEUE",
-		"BUILDKITE_GHA_PLAN_DIGEST",
-	} {
-		t.Setenv(name, "")
-	}
-}
-
-func publishedCLIManifest(t *testing.T, runner *cliCaptureRunner, job plan.Job, planDigest string) transport.ResultManifest {
-	t.Helper()
-	path := transport.ResultPath(job.Target.StepKey, planDigest)
-	data, ok := runner.uploaded[path]
-	if !ok {
-		t.Fatalf("authoritative result %q was not uploaded; uploads = %#v", path, runner.uploaded)
-	}
-	producer := transport.Producer{BuildID: cliTestBuildID, JobID: cliTestJobID, StepKey: job.Target.StepKey}
-	manifest, err := transport.VerifyResultManifest(data, planDigest, producer)
-	if err != nil {
-		t.Fatalf("verify published result: %v\n%s", err, data)
-	}
-	return manifest
 }
